Struct gdk::app_launch_context::AppLaunchContext

pub struct AppLaunchContext(_, _);

Application launching — startup notification for applications.

Methods

impl AppLaunchContext

fn new() -> AppLaunchContext

Creates a new AppLaunchContext.

fn set_display(&self, display: &Display)

Sets the display on which applications will be launched when using this context. See also AppLaunchContext::set_screen().

fn set_screen(&self, screen: &Screen)

Sets the screen on which applications will be launched when using this context. See also AppLaunchContext::set_display().

If both screen and display are set, the screen takes priority. If neither screen or display are set, the default screen and display are used.

fn set_desktop(&self, desktop: i32)

Sets the workspace on which applications will be launched when using this context when running under a window manager that supports multiple workspaces, as described in the Extended Window Manager Hints.

When the workspace is not specified or desktop is set to -1, it is up to the window manager to pick one, typically it will be the current workspace.

fn set_timestamp(&self, timestamp: u32)

Sets the timestamp of self. The timestamp should ideally be taken from the event that triggered the launch.

Window managers can use this information to avoid moving the focus to the newly launched application when the user is busy typing in another window. This is also known as 'focus stealing prevention'.

fn set_icon_name(&self, icon_name: &str)

Sets the icon for applications that are launched with this context. The icon_name will be interpreted in the same way as the Icon field in desktop files. See also AppLaunchContext::set_icon().

If both icon and icon_name are set, the icon_name takes priority. If neither icon or icon_name is set, the icon is taken from either the file that is passed to launched application or from the GAppInfo for the launched application itself.