Trait gdk::cairo_interaction::ContextExt

pub trait ContextExt {
    fn create_from_window(window: &Window) -> Context;
    fn get_clip_rectangle(&self) -> Option<RectangleInt>;
    fn set_source_rgba(&self, rgba: &GdkRGBA);
    fn set_source_pixbuf(&self, pixbuf: &Pixbuf, x: f64, y: f64);
    fn set_source_window(&self, window: &Window, x: f64, y: f64);
    fn rectangle(&self, rectangle: &RectangleInt);
}

Required Methods

fn create_from_window(window: &Window) -> Context

Creates a Cairo context for drawing to window.

Note that calling reset_clip() on the resulting Context will produce undefined results, so avoid it at all costs.

fn get_clip_rectangle(&self) -> Option<RectangleInt>

This is a convenience function around clip_extents(). It rounds the clip extents to integer coordinates and returns a RectangleInt, or None if no clip area exists.

fn set_source_rgba(&self, rgba: &GdkRGBA)

Sets the specified GdkRGBA as the source color of cr.

fn set_source_pixbuf(&self, pixbuf: &Pixbuf, x: f64, y: f64)

Sets the given pixbuf as the source pattern for cr. The pattern has an extend mode of ExtendNone and is aligned so that the origin of pixbuf is (x, y).

fn set_source_window(&self, window: &Window, x: f64, y: f64)

Sets the given window as the source pattern for cr. The pattern has an extend mode of ExtendNone and is aligned so that the origin of window is (x , y). The window contains all its subwindows when rendering.

Note that the contents of window are undefined outside of the visible part of window, so use this function with care.

fn rectangle(&self, rectangle: &RectangleInt)

Adds the given rectangle to the current path of cr.