Struct gdk::cursor::Cursor

pub struct Cursor(_, _);

Standard and pixmap cursors.

Methods

impl Cursor

fn new(cursor_type: Type) -> Cursor

Creates a new cursor from the set of builtin cursors for the default display. See Cursor::new_for_display().

To make the cursor invisible, use Type::BlankCursor.

Warning : Cursor::new() has been deprecated since version 3.16 and should not be used in newly-written code.

Use Cursor::new_for_display() instead.

fn new_from_pixbuf(display: &Display, pixbuf: &Pixbuf, x: i32, y: i32) -> Cursor

Creates a new cursor from a pixbuf.

Not all GDK backends support RGBA cursors. If they are not supported, a monochrome approximation will be displayed. The functions Display::supports_cursor_alpha() and Display::supports_cursor_color() can be used to determine whether RGBA cursors are supported; Display::get_default_cursor_size() and Display::get_maximal_cursor_size() give information about cursor sizes.

If x or y are -1, the pixbuf must have options named “x_hot” and “y_hot”, resp., containing integer values between 0 and the width resp. height of the pixbuf. (Since: 3.0)

On the X backend, support for RGBA cursors requires a sufficently new version of the X Render extension.

fn new_from_name(display: &Display, name: &str) -> Option<Cursor>

Creates a new cursor by looking up name in the current cursor theme.

fn new_for_display(display: &Display, cursor_type: Type) -> Cursor

Creates a new cursor from the set of builtin cursors.

You can find a list of useful ones here: https://developer.gnome.org/gdk3/stable/gdk3-Cursors.html#gdk-cursor-new-for-display

fn get_display(&self) -> Display

Returns the display on which the Cursor is defined.

fn get_image(&self) -> Option<Pixbuf>

Returns a Pixbuf with the image used to display the cursor.

Note that depending on the capabilities of the windowing system and on the cursor, GDK may not be able to obtain the image data. In this case, None is returned.

fn get_cursor_type(&self) -> Type

Returns the cursor type for this cursor.