Struct gdk::device::Device

pub struct Device(_, _);

Object representing an input device.

Methods

impl Device

fn get_name(&self) -> Option<String>

Determines the name of the device.

fn get_vendor_id(&self) -> Option<String>

Returns the vendor ID of this device, or NULL if this information couldn't be obtained. This ID is retrieved from the device, and is thus constant for it.

This function, together with gdk_device_get_product_id(), can be used to eg. compose GSettings paths to store settings for this device.

fn get_device_settings(device: &Device) -> GSettings {
    GSettings *settings;
    GdkDevice *device;
    gchar *path;
 
    let vendor = device.get_vendor_id().unwrap();
    let product = device.get_product_id().unwrap();
 
    let path = format!("/org/example/app/devices/{}:{}/", vendor, product);
    g_settings_new_with_path(DEVICE_SCHEMA, path);
}

fn get_product_id(&self) -> Option<String>

Returns the product ID of this device, or None if this information couldn't be obtained. This ID is retrieved from the device, and is thus constant for it. See Device::get_vendor_id() for more information.

fn get_source(&self) -> InputSource

Determines the type of the device.

fn set_mode(&self, mode: InputMode)

Sets a the mode of an input device. The mode controls if the device is active and whether the device’s range is mapped to the entire screen or to a single window.

Note: This is only meaningful for floating devices, master devices (and slaves connected to these) drive the pointer cursor, which is not limited by the input mode.

fn get_mode(&self) -> InputMode

Determines the mode of the device.

fn set_key(&self, index_: u32, keyval: u32, modifiers: ModifierType)

Specifies the X key event to generate when a macro button of a device is pressed.

fn get_key(&self, index_: u32, keyval: &mut u32, modifiers: &mut ModifierType) -> bool

If index_ has a valid keyval, this function will return true and fill in keyval and modifiers with the keyval settings.

fn set_axis_use(&self, index_: u32, use_: AxisUse)

Specifies how an axis of a device is used.

fn get_axis_use(&self, index_: u32) -> AxisUse

Returns the axis use for index_.

fn get_associated_device(&self) -> Option<Device>

Returns the associated self to self , if self is of type DeviceType::Master, it will return the paired pointer or keyboard.

If self is of type DeviceType::Slave, it will return the master device to which self is attached to.

If self is of type DeviceType::Floating, None will be returned, as there is no associated device.

fn get_device_type(&self) -> Type

Returns the device type for self.

fn get_display(&self) -> Display

Returns the Display to which self pertains.

fn get_has_cursor(&self) -> bool

Determines whether the pointer follows device motion. This is not meaningful for keyboard devices, which don't have a pointer.

fn get_n_axes(&self) -> i32

Returns the number of axes the device currently has.

fn get_n_keys(&self) -> i32

Returns the number of keys the device currently has.

fn warp(&self, screen: &Screen, x: i32, y: i32)

Warps self in display to the point x, y on the screen screen, unless the device is confined to a window by a grab, in which case it will be moved as far as allowed by the grab. Warping the pointer creates events as if the user had moved the mouse instantaneously to the destination.

Note that the pointer should normally be under the control of the user. This function was added to cover some rare use cases like keyboard navigation support for the color picker in the gtk::ColorSelectionDialog.

fn grab(&self, window: &Window, grab_ownership: GrabOwnership, owner_events: bool, event_mask: EventMask, cursor: &Cursor, time_: u32) -> GrabStatus

Grabs the device so that all events coming from this device are passed to this application until the device is ungrabbed with Device::ungrab(), or the window becomes unviewable. This overrides any previous grab on the device by this client.

Note that self and window need to be on the same display.

Device grabs are used for operations which need complete control over the given device events (either pointer or keyboard). For example in GTK+ this is used for Drag and Drop operations, popup menus and such.

Note that if the event mask of an X window has selected both button press and button release events, then a button press event will cause an automatic pointer grab until the button is released. X does this automatically since most applications expect to receive button press and release events in pairs. It is equivalent to a pointer grab on the window with owner_events set to true.

If you set up anything at the time you take the grab that needs to be cleaned up when the grab ends, you should handle the GdkEventGrabBroken events that are emitted when the grab ends unvoluntarily.

fn ungrab(&self, time_: u32)

Release any grab on self.

fn get_axis(&self, axes: &mut [f64], use_: AxisUse, value: &mut f64) -> bool

Interprets an array of double as axis values for a given device, and locates the value in the array for a given axis use.