Struct gdk::display::Display

pub struct Display(_, _);

Controls a set of Screens and their associated input devices.

Display objects purpose are two fold: * To manage and provide information about input devices (pointers and keyboards) * To manage and provide information about the available Screens

Display objects are the GDK representation of an X Display, which can be described as a workstation consisting of a keyboard, a pointing device (such as a mouse) and one or more screens. It is used to open and keep track of various GdkScreen objects currently instantiated by the application. It is also used to access the keyboard(s) and mouse pointer(s) of the display.

Most of the input device handling has been factored out into the separate DeviceManager object. Every display has a device manager, which you can obtain using Display::get_device_manager().

Methods

impl Display

fn open(display_name: &str) -> Option<Display>

Opens a display.

fn get_default() -> Option<Display>

Gets the default Display. This is a convenience function for: DisplayManager::get_default_display(DisplayManager::get()).

fn get_name(&self) -> String

Gets the name of the display.

fn get_screen(&self, screen_num: i32) -> Screen

Returns a screen object for one of the screens of the display.

fn get_default_screen(&self) -> Screen

Get the default Screen for self.

fn get_device_manager(&self) -> Option<DeviceManager>

Returns the DeviceManager associated to self.

fn device_is_grabbed(&self, device: &Device) -> bool

Returns true if there is an ongoing grab on device for self.

fn beep(&self)

Emits a short beep on self.

fn sync(&self)

Flushes any requests queued for the windowing system and waits until all requests have been handled. This is often used for making sure that the display is synchronized with the current state of the program. Calling Display::sync() before gdk::error_trap_pop() makes sure that any errors generated from earlier requests are handled before the error trap is removed.

This is most useful for X11. On windowing systems where requests are handled synchronously, this function will do nothing.

fn flush(&self)

Flushes any requests queued for the windowing system; this happens automatically when the main loop blocks waiting for new events, but if your application is drawing without returning control to the main loop, you may need to call this function explicitly. A common case where this function needs to be called is when an application is executing drawing commands from a thread other than the thread where the main loop is running.

This is most useful for X11. On windowing systems where requests are handled synchronously, this function will do nothing.

fn close(&self)

Closes the connection to the windowing system for the given display, and cleans up associated resources.

fn is_closed(&self) -> bool

Finds out if the display has been closed.

fn has_pending(&self) -> bool

Returns whether the display has events that are waiting to be processed.

fn set_double_click_time(&self, msec: u32)

Sets the double click time (two clicks within this time interval count as a double click and result in a GDK_2BUTTON_PRESS event). Applications should not set this, it is a global user-configured setting.

fn set_double_click_distance(&self, msec: u32)

Sets the double click distance (two clicks within this distance count as a double click and result in a GDK_2BUTTON_PRESS event). See also Display::set_double_click_time(). Applications should not set this, it is a global user-configured setting.

fn supports_cursor_color(&self) -> bool

Returns true if multicolored cursors are supported on self. Otherwise, cursors have only a forground and a background color.

fn supports_cursor_alpha(&self) -> bool

Returns true if cursors can use an 8bit alpha channel on self. Otherwise, cursors are restricted to bilevel alpha (i.e. a mask).

fn get_default_cursor_size(&self) -> u32

Returns the default size to use for cursors on self.

fn get_maximal_cursor_size(&self, width: &mut u32, height: &mut u32)

Gets the maximal size to use for cursors on self.

fn get_default_group(&self) -> Window

Returns the default group leader window for all toplevel windows on self. This window is implicitly created by GDK. See Window::set_group().

fn supports_selection_notification(&self) -> bool

Returns whether EventOwnerChange events will be sent when the owner of a selection changes.

fn request_selection_notification(&self, selection: &Atom) -> bool

Request EventOwnerChange events for ownership changes of the selection named by the given atom.

fn supports_clipboard_persistence(&self) -> bool

Returns whether the speicifed display supports clipboard persistance; i.e. if it’s possible to store the clipboard data after an application has quit. On X11 this checks if a clipboard daemon is running.

fn supports_shapes(&self) -> bool

Returns true if Window::shape_combine_mask() can be used to create shaped windows on self.

fn supports_input_shapes(&self) -> bool

Returns true if Window::input_shape_combine_mask() can be used to modify the input shape of windows on self.

fn supports_composite(&self) -> bool

Returns true if Window::set_composited() can be used to redirect drawing on the window using compositing.

Currently this only works on X11 with XComposite and XDamage extensions available.

Deprecated since GDK 3.16

fn get_app_launch_context(&self) -> AppLaunchContext

Returns a AppLaunchContext suitable for launching applications on the given display.

fn notify_startup_complete(&self, startup_id: &str)

Indicates to the GUI environment that the application has finished loading, using a given identifier.

GTK+ will call this function automatically for GtkWindow with custom startup-notification identifier unless gtk::Window::set_auto_startup_notification() is called to disable that feature.