Enum gdk::CursorType

pub enum CursorType {
    XCursor,
    Arrow,
    BasedArrowDown,
    BasedArrowUp,
    Boat,
    Bogosity,
    BottomLeftCorner,
    BottomRightCorner,
    BottomSide,
    BottomTee,
    BoxSpiral,
    CenterPtr,
    Circle,
    Clock,
    CoffeeMug,
    Cross,
    CrossReverse,
    Crosshair,
    DiamondCross,
    Dot,
    Dotbox,
    DoubleArrow,
    DraftLarge,
    DraftSmall,
    DrapedBox,
    Exchange,
    Fleur,
    Gobbler,
    Gumby,
    Hand1,
    Hand2,
    Heart,
    Icon,
    IronCross,
    LeftPtr,
    LeftSide,
    LeftTee,
    Leftbutton,
    LlAngle,
    LrAngle,
    Man,
    Middlebutton,
    Mouse,
    Pencil,
    Pirate,
    Plus,
    QuestionArrow,
    RightPtr,
    RightSide,
    RightTee,
    Rightbutton,
    RtlLogo,
    Sailboat,
    SbDownArrow,
    SbHDoubleArrow,
    SbLeftArrow,
    SbRightArrow,
    SbUpArrow,
    SbVDoubleArrow,
    Shuttle,
    Sizing,
    Spider,
    Spraycan,
    Star,
    Target,
    Tcross,
    TopLeftArrow,
    TopLeftCorner,
    TopRightCorner,
    TopSide,
    TopTee,
    Trek,
    UlAngle,
    Umbrella,
    UrAngle,
    Watch,
    Xterm,
    LastCursor,
    BlankCursor,
    CursorIsPixmap,
    // some variants omitted
}

Predefined cursors.

Note that these IDs are directly taken from the X cursor font, and many of these cursors are either not useful, or are not available on other platforms.

The recommended way to create cursors is to use Cursor::new_from_name.

Variants
XCursorArrowBasedArrowDownBasedArrowUpBoatBogosityBottomLeftCornerBottomRightCornerBottomSideBottomTeeBoxSpiralCenterPtrCircleClockCoffeeMugCrossCrossReverseCrosshairDiamondCrossDotDotboxDoubleArrowDraftLargeDraftSmallDrapedBoxExchangeFleurGobblerGumbyHand1Hand2HeartIconIronCrossLeftPtrLeftSideLeftTeeLeftbuttonLlAngleLrAngleManMiddlebuttonMousePencilPiratePlusQuestionArrowRightPtrRightSideRightTeeRightbuttonRtlLogoSailboatSbDownArrowSbHDoubleArrowSbLeftArrowSbRightArrowSbUpArrowSbVDoubleArrowShuttleSizingSpiderSpraycanStarTargetTcrossTopLeftArrowTopLeftCornerTopRightCornerTopSideTopTeeTrekUlAngleUmbrellaUrAngleWatchXtermLastCursorBlankCursorCursorIsPixmap