Enum gdk::ModifierIntent

#[non_exhaustive]
pub enum ModifierIntent {
    PrimaryAccelerator,
    ContextMenu,
    ExtendSelection,
    ModifySelection,
    NoTextInput,
    ShiftGroup,
    DefaultModMask,
    // some variants omitted
}

This enum is used with Keymap::modifier_mask() in order to determine what modifiers the currently used windowing system backend uses for particular purposes. For example, on X11/Windows, the Control key is used for invoking menu shortcuts (accelerators), whereas on Apple computers it’s the Command key (which correspond to ModifierType::CONTROL_MASK and ModifierType::MOD2_MASK, respectively).

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

PrimaryAccelerator

the primary modifier used to invoke menu accelerators.

ContextMenu

the modifier used to invoke context menus. Note that mouse button 3 always triggers context menus. When this modifier is not 0, it additionally triggers context menus when used with mouse button 1.

ExtendSelection

the modifier used to extend selections using modifier-click or modifier-cursor-key

ModifySelection

the modifier used to modify selections, which in most cases means toggling the clicked item into or out of the selection.

NoTextInput

when any of these modifiers is pressed, the key event cannot produce a symbol directly. This is meant to be used for input methods, and for use cases like typeahead search.

ShiftGroup

the modifier that switches between keyboard groups (AltGr on X11/Windows and Option/Alt on OS X).

DefaultModMask

The set of modifier masks accepted as modifiers in accelerators. Needed because Command is mapped to MOD2 on OSX, which is widely used, but on X11 MOD2 is NumLock and using that for a mod key is problematic at best. Ref: https://bugzilla.gnome.org/show_bug.cgi?id=736125.