Enum gdk::WindowWindowClass

pub enum WindowWindowClass {
    InputOutput,
    InputOnly,
}

WindowWindowClass::InputOutput windows are the standard kind of window you might expect. Such windows receive events and are also displayed on screen. WindowWindowClass::InputOnly windows are invisible; they are usually placed above other windows in order to trap or filter the events. You can’t draw on WindowWindowClass::InputOnly windows.

Variants

InputOutput

window for graphics and events

InputOnly

window for events only