Module gdk::enums::modifier_type

Constants

Button1Mask
the first mouse button.

Button2Mask
the second mouse button.

Button3Mask
the third mouse button.

Button4Mask
the fourth mouse button.

Button5Mask
the fifth mouse button.

ControlMask
the Control key.

HyperMask
the Hyper modifier.

LockMask
a Lock key (depending on the modifier mapping of the X server this may either be CapsLock or ShiftLock).

MetaMask
the Meta modifier.

Mod1Mask
the fourth modifier key (it depends on the modifier mapping of the X server which key is interpreted as this modifier, but normally it is the Alt key).

Mod2Mask
the fifth modifier key (it depends on the modifier mapping of the X server which key is interpreted as this modifier).

Mod3Mask
the sixth modifier key (it depends on the modifier mapping of the X server which key is interpreted as this modifier).

Mod4Mask
the seventh modifier key (it depends on the modifier mapping of the X server which key is interpreted as this modifier).

Mod5Mask
the eighth modifier key (it depends on the modifier mapping of the X server which key is interpreted as this modifier).

ModifierMask
a mask covering all modifier types.

ReleaseMask
not used in GDK itself. GTK+ uses it to differentiate between (keyval, modifiers) pairs from key press and release events.

ShiftMask
the Shift key.,

SuperMask
the Super modifier.