Function gdk::error_trap_pop

pub fn error_trap_pop() -> i32

Removes an error trap pushed with error_trap_push(). May block until an error has been definitively received or not received from the X server. error_trap_pop_ignored() is preferred if you don’t need to know whether an error occurred, because it never has to block. If you don’t need the return value of error_trap_pop(), use error_trap_pop_ignored().

Prior to GDK 3.0, this function would not automatically sync for you, so you had to flush() if your last call to Xlib was not a blocking round trip.

Returns

X error code or 0 on success