Function gdk::notify_startup_complete_with_id

pub fn notify_startup_complete_with_id(startup_id: &str)

Indicates to the GUI environment that the application has finished loading, using a given identifier.

GTK+ will call this function automatically for GtkWindow with custom startup-notification identifier unless gtk_window_set_auto_startup_notification() is called to disable that feature.

startup_id

a startup-notification identifier, for which notification process should be completed