gdk

Function selection_owner_get

pub fn selection_owner_get(selection: &Atom) -> Option<Window>

Determines the owner of the given selection.

selection

an atom indentifying a selection.

Returns

if there is a selection owner for this window, and it is a window known to the current process, the Window that owns the selection, otherwise None. Note that the return value may be owned by a different process if a foreign window was previously created for that window, but a new foreign window will never be created by this call.