Function gdk::selection_owner_get_for_display

pub fn selection_owner_get_for_display(
    display: &Display,
    selection: &Atom
) -> Option<Window>

Determine the owner of the given selection.

Note that the return value may be owned by a different process if a foreign window was previously created for that window, but a new foreign window will never be created by this call.

display

a Display

selection

an atom indentifying a selection

Returns

if there is a selection owner for this window, and it is a window known to the current process, the Window that owns the selection, otherwise None.