Function gdk::functions::query_depths

pub fn query_depths() -> Vec<i32>

This function returns the available bit depths for the default screen. It’s equivalent to listing the visuals (list_visuals()) and then looking at the depth field in each visual, removing duplicates.

The array returned by this function should not be freed.

Deprecated since 3.22

Visual selection should be done using Screen::system_visual() and Screen::rgba_visual()

Returns

depths

return location for available depths