Module gdk::pixbuf

Reexports

pub use self::animation::{PixbufAnimation, PixbufAnimationIter, PixbufSimpleAnim};

pub use self::format::PixbufFormat;

pub use self::loader::PixbufLoader;

Modules

animation
Animations — Animated images.

format
A GdkPixbufFormat contains information about the image format accepted by a module. Only modules should access the fields directly, applications should use the gdk_pixbuf_format_* functions.

loader

Structs

Pixbuf
Represents an image.