gdk::pixbuf::Pixbuf

type Pixbuf = Object<GdkPixbuf>;

This is the main structure in the &gdk-pixbuf; library. It is used to represent images. It contains information about the image's pixel data, its color space, bits per sample, width and height, and the rowstride (the number of bytes between the start of one row and the start of the next).