Trait gdk::prelude::DeviceManagerExt

pub trait DeviceManagerExt {
    fn get_client_pointer(&self) -> Option<Device>;

    fn get_display(&self) -> Option<Display>;

    fn list_devices(&self, type_: DeviceType) -> Vec<Device>;

    fn connect_device_added<F: Fn(&Self, &Device) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_device_changed<F: Fn(&Self, &Device) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_device_removed<F: Fn(&Self, &Device) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_display_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all DeviceManager methods.

Implementors

DeviceManager

Required Methods

fn get_client_pointer(&self) -> Option<Device>

Returns the client pointer, that is, the master pointer that acts as the core pointer for this application. In X11, window managers may change this depending on the interaction pattern under the presence of several pointers.

You should use this function seldomly, only in code that isn’t triggered by a GdkEvent and there aren’t other means to get a meaningful Device to operate on.

Deprecated since 3.20

Use SeatExt::get_pointer instead.

Returns

The client pointer. This memory is owned by GDK and must not be freed or unreferenced.

fn get_display(&self) -> Option<Display>

Gets the Display associated to self.

Returns

the Display to which self is associated to, or None. This memory is owned by GDK and must not be freed or unreferenced.

fn list_devices(&self, type_: DeviceType) -> Vec<Device>

Returns the list of devices of type type_ currently attached to self.

Deprecated since 3.20

, use SeatExt::get_pointer, SeatExt::get_keyboard and SeatExt::get_slaves instead.

type_

device type to get.

Returns

a list of GdkDevices. The returned list must be freed with g_list_free (). The list elements are owned by GTK+ and must not be freed or unreffed.

fn connect_device_added<F: Fn(&Self, &Device) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::device-added signal is emitted either when a new master pointer is created, or when a slave (Hardware) input device is plugged in.

device

the newly added Device.

fn connect_device_changed<F: Fn(&Self, &Device) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::device-changed signal is emitted whenever a device has changed in the hierarchy, either slave devices being disconnected from their master device or connected to another one, or master devices being added or removed a slave device.

If a slave device is detached from all master devices (DeviceExt::get_associated_device returns None), its DeviceType will change to DeviceType::Floating, if it's attached, it will change to DeviceType::Slave.

device

the Device that changed.

fn connect_device_removed<F: Fn(&Self, &Device) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::device-removed signal is emitted either when a master pointer is removed, or when a slave (Hardware) input device is unplugged.

device

the just removed Device.

fn connect_property_display_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId