Trait gdk::prelude::DragContextExt

pub trait DragContextExt {
    fn get_actions(&self) -> DragAction;

    fn get_dest_window(&self) -> Window;

    fn get_device(&self) -> Device;

    fn get_drag_window(&self) -> Option<Window>;

    fn get_protocol(&self) -> DragProtocol;

    fn get_selected_action(&self) -> DragAction;

    fn get_source_window(&self) -> Window;

    fn get_suggested_action(&self) -> DragAction;

    fn list_targets(&self) -> Vec<Atom>;

    fn manage_dnd(&self, ipc_window: &Window, actions: DragAction) -> bool;

    fn set_device<P: IsA<Device>>(&self, device: &P);

    fn set_hotspot(&self, hot_x: i32, hot_y: i32);

    fn connect_action_changed<F: Fn(&Self, DragAction) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_cancel<F: Fn(&Self, DragCancelReason) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_dnd_finished<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_drop_performed<F: Fn(&Self, i32) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all DragContext methods.

Implementors

DragContext

Required Methods

fn get_actions(&self) -> DragAction

Determines the bitmask of actions proposed by the source if DragContextExt::get_suggested_action returns DragAction::Ask.

Returns

the DragAction flags

fn get_dest_window(&self) -> Window

Returns the destination windw for the DND operation.

Returns

a Window

fn get_device(&self) -> Device

Returns the Device associated to the drag context.

Returns

The Device associated to self.

fn get_drag_window(&self) -> Option<Window>

Returns the window on which the drag icon should be rendered during the drag operation. Note that the window may not be available until the drag operation has begun. GDK will move the window in accordance with the ongoing drag operation. The window is owned by self and will be destroyed when the drag operation is over.

Feature: v3_20

Returns

the drag window, or None

fn get_protocol(&self) -> DragProtocol

Returns the drag protocol thats used by this context.

Returns

the drag protocol

fn get_selected_action(&self) -> DragAction

Determines the action chosen by the drag destination.

Returns

a DragAction value

fn get_source_window(&self) -> Window

Returns the Window where the DND operation started.

Returns

a Window

fn get_suggested_action(&self) -> DragAction

Determines the suggested drag action of the context.

Returns

a DragAction value

fn list_targets(&self) -> Vec<Atom>

Retrieves the list of targets of the context.

Returns

a glib::List of targets

fn manage_dnd(&self, ipc_window: &Window, actions: DragAction) -> bool

Requests the drag and drop operation to be managed by self. When a drag and drop operation becomes managed, the DragContext will internally handle all input and source-side EventDND events as required by the windowing system.

Once the drag and drop operation is managed, the drag context will emit the following signals: - The DragContext::action-changed signal whenever the final action to be performed by the drag and drop operation changes. - The DragContext::drop-performed signal after the user performs the drag and drop gesture (typically by releasing the mouse button). - The DragContext::dnd-finished signal after the drag and drop operation concludes (after all GdkSelection transfers happen). - The DragContext::cancel signal if the drag and drop operation is finished but doesn't happen over an accepting destination, or is cancelled through other means.

Feature: v3_20

ipc_window

Window to use for IPC messaging/events

actions

the actions supported by the drag source

Returns

true if the drag and drop operation is managed.

fn set_device<P: IsA<Device>>(&self, device: &P)

Associates a Device to self, so all Drag and Drop events for self are emitted as if they came from this device.

device

a Device

fn set_hotspot(&self, hot_x: i32, hot_y: i32)

Sets the position of the drag window that will be kept under the cursor hotspot. Initially, the hotspot is at the top left corner of the drag window.

Feature: v3_20

hot_x

x coordinate of the drag window hotspot

hot_y

y coordinate of the drag window hotspot

fn connect_action_changed<F: Fn(&Self, DragAction) + 'static>(
&self,
f: F
) -> SignalHandlerId

A new action is being chosen for the drag and drop operation.

This signal will only be emitted if the DragContext manages the drag and drop operation. See DragContextExt::manage_dnd for more information.

Feature: v3_20

action

The action currently chosen

fn connect_cancel<F: Fn(&Self, DragCancelReason) + 'static>(
&self,
f: F
) -> SignalHandlerId

The drag and drop operation was cancelled.

This signal will only be emitted if the DragContext manages the drag and drop operation. See DragContextExt::manage_dnd for more information.

Feature: v3_20

reason

The reason the context was cancelled

fn connect_dnd_finished<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

The drag and drop operation was finished, the drag destination finished reading all data. The drag source can now free all miscellaneous data.

This signal will only be emitted if the DragContext manages the drag and drop operation. See DragContextExt::manage_dnd for more information.

Feature: v3_20

fn connect_drop_performed<F: Fn(&Self, i32) + 'static>(
&self,
f: F
) -> SignalHandlerId

The drag and drop operation was performed on an accepting client.

This signal will only be emitted if the DragContext manages the drag and drop operation. See DragContextExt::manage_dnd for more information.

Feature: v3_20

time

the time at which the drop happened.