Trait gdk::prelude::DrawingContextExt

pub trait DrawingContextExt {
    fn get_cairo_context(&self) -> Option<Context>;

    fn get_clip(&self) -> Option<Region>;

    fn get_window(&self) -> Option<Window>;

    fn is_valid(&self) -> bool;

    fn connect_property_clip_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_window_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all DrawingContext methods.

Feature: v3_22

Implementors

DrawingContext

Required Methods

fn get_cairo_context(&self) -> Option<Context>

Retrieves a Cairo context to be used to draw on the Window that created the DrawingContext.

The returned context is guaranteed to be valid as long as the DrawingContext is valid, that is between a call to WindowExt::begin_draw_frame and WindowExt::end_draw_frame.

Feature: v3_22

Returns

a Cairo context to be used to draw the contents of the Window. The context is owned by the DrawingContext and should not be destroyed

fn get_clip(&self) -> Option<Region>

Retrieves a copy of the clip region used when creating the self.

Feature: v3_22

Returns

a Cairo region

fn get_window(&self) -> Option<Window>

Retrieves the window that created the drawing self.

Feature: v3_22

Returns

a Window

fn is_valid(&self) -> bool

Checks whether the given DrawingContext is valid.

Feature: v3_22

Returns

true if the context is valid

fn connect_property_clip_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_window_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId