Trait gdk::prelude::WindowExtManual

pub trait WindowExtManual: 'static {
    unsafe fn set_user_data<T>(&self, user_data: &mut T);

    unsafe fn get_user_data<T>(&self) -> &mut T;

    fn get_default_root_window() -> Window;

    fn offscreen_window_set_embedder(&self, embedder: &Window);

    fn offscreen_window_get_embedder(&self) -> Option<Window>;

    fn offscreen_window_get_surface(&self) -> Option<Surface>;

    fn get_pixbuf(
        &self, 
        src_x: i32, 
        src_y: i32, 
        width: i32, 
        height: i32
    ) -> Option<Pixbuf>;

    fn get_background_pattern(&self) -> Option<Pattern>;

    fn set_background_pattern(&self, pattern: Option<&Pattern>);
}

Required methods

unsafe fn set_user_data<T>(&self, user_data: &mut T)

For most purposes this function is deprecated in favor of gobject::Object::set_data. However, for historical reasons GTK+ stores the GtkWidget that owns a Window as user data on the Window. So, custom widget implementations should use this function for that. If GTK+ receives an event for a Window, and the user data for the window is non-None, GTK+ will assume the user data is a GtkWidget, and forward the event to that widget.

user_data

user data

unsafe fn get_user_data<T>(&self) -> &mut T

Retrieves the user data for self, which is normally the widget that self belongs to. See Window::set_user_data.

data

return location for user data

fn get_default_root_window() -> Window

fn offscreen_window_set_embedder(&self, embedder: &Window)

fn offscreen_window_get_embedder(&self) -> Option<Window>

fn offscreen_window_get_surface(&self) -> Option<Surface>

fn get_pixbuf(
&self,
src_x: i32,
src_y: i32,
width: i32,
height: i32
) -> Option<Pixbuf>

fn get_background_pattern(&self) -> Option<Pattern>

Gets the pattern used to clear the background on self.

Deprecated since 3.22

Don't use this function

Returns

The pattern to use for the background or None if there is no background.

fn set_background_pattern(&self, pattern: Option<&Pattern>)

Sets the background of self.

A background of None means that the window won't have any background. On the X11 backend it's also possible to inherit the background from the parent window using gdk_x11_get_parent_relative_pattern.

The windowing system will normally fill a window with its background when the window is obscured then exposed.

Deprecated since 3.22

Don't use this function

pattern

a pattern to use, or None

Loading content...