Struct gdk::AppLaunchContext

pub struct AppLaunchContext(_, _);

AppLaunchContext is an implementation of gio::AppLaunchContext that handles launching an application in a graphical context. It provides startup notification and allows to launch applications on a specific screen or workspace.

Launching an application

GdkAppLaunchContext *context;

context = gdk_display_get_app_launch_context (display);

gdk_app_launch_context_set_screen (screen);
gdk_app_launch_context_set_timestamp (event->time);

if (!g_app_info_launch_default_for_uri ("http://www.gtk.org", context, &error))
  g_warning ("Launching failed: %s\n", error->message);

g_object_unref (context);

Methods

impl AppLaunchContext

fn new() -> AppLaunchContext

Creates a new AppLaunchContext.

Deprecated since 3.0

Use Display::get_app_launch_context instead

Returns

a new AppLaunchContext

fn set_display(&self, display: &Display)

Sets the display on which applications will be launched when using this context. See also AppLaunchContext::set_screen.

Deprecated since 3.0

Use Display::get_app_launch_context instead

display

a Display

fn set_screen(&self, screen: &Screen)

Sets the screen on which applications will be launched when using this context. See also AppLaunchContext::set_display.

If both screen and display are set, the screen takes priority. If neither screen or display are set, the default screen and display are used.

screen

a Screen

fn set_desktop(&self, desktop: i32)

Sets the workspace on which applications will be launched when using this context when running under a window manager that supports multiple workspaces, as described in the Extended Window Manager Hints.

When the workspace is not specified or desktop is set to -1, it is up to the window manager to pick one, typically it will be the current workspace.

desktop

the number of a workspace, or -1

fn set_timestamp(&self, timestamp: u32)

Sets the timestamp of self. The timestamp should ideally be taken from the event that triggered the launch.

Window managers can use this information to avoid moving the focus to the newly launched application when the user is busy typing in another window. This is also known as 'focus stealing prevention'.

timestamp

a timestamp

fn set_icon_name(&self, icon_name: &str)

Sets the icon for applications that are launched with this context. The icon_name will be interpreted in the same way as the Icon field in desktop files. See also AppLaunchContext::set_icon.

If both icon and icon_name are set, the icon_name takes priority. If neither icon or icon_name is set, the icon is taken from either the file that is passed to launched application or from the gio::AppInfo for the launched application itself.

icon_name

an icon name, or None