gdk

Struct Atom

pub struct Atom(/* private fields */);

An opaque type representing a string as an index into a table of strings on the X server.

Implementations

impl Atom

pub fn intern(atom_name: &str) -> Atom

Finds or creates an atom corresponding to a given string.

atom_name

a string.

only_if_exists

if true, GDK is allowed to not create a new atom, but just return GDK_NONE if the requested atom doesn’t already exists. Currently, the flag is ignored, since checking the existance of an atom is as expensive as creating it.

Returns

the atom corresponding to atom_name.

pub fn name(self) -> GString

Determines the string corresponding to an atom.

Returns

a newly-allocated string containing the string corresponding to self. When you are done with the return value, you should free it using g_free().

pub fn value(self) -> usize