Struct gdk::DisplayManager

pub struct DisplayManager(_, _);

The purpose of the DisplayManager singleton object is to offer notification when displays appear or disappear or the default display changes.

You can use DisplayManager::get to obtain the DisplayManager singleton, but that should be rarely necessary. Typically, initializing GTK+ opens a display that you can work with without ever accessing the DisplayManager.

The GDK library can be built with support for multiple backends. The DisplayManager object determines which backend is used at runtime.

When writing backend-specific code that is supposed to work with multiple GDK backends, you have to consider both compile time and runtime. At compile time, use the GDK_WINDOWING_X11, GDK_WINDOWING_WIN32 macros, etc. to find out which backends are present in the GDK library you are building your application against. At runtime, use type-check macros like GDK_IS_X11_DISPLAY() to find out which backend is in use:

Backend-specific code ## {backend-specific}

#ifdef GDK_WINDOWING_X11
  if (GDK_IS_X11_DISPLAY (display))
    {
      // make X11-specific calls here
    }
  else
#endif
#ifdef GDK_WINDOWING_QUARTZ
  if (GDK_IS_QUARTZ_DISPLAY (display))
    {
      // make Quartz-specific calls here
    }
  else
#endif
  g_error ("Unsupported GDK backend");

Methods

impl DisplayManager

fn get_default_display(&self) -> Option<Display>

Gets the default Display.

Returns

a Display, or None if there is no default display.

fn list_displays(&self) -> Vec<Display>

List all currently open displays.

Returns

a newly allocated glib::SList of Display objects. Free with glib::SList::free when you are done with it.

fn open_display(&self, name: &str) -> Option<Display>

Opens a display.

name

the name of the display to open

Returns

a Display, or None if the display could not be opened

fn set_default_display(&self, display: &Display)

Sets display as the default display.

display

a Display

fn get_property_default_display(&self) -> Option<Display>

fn set_property_default_display(&self, default_display: Option<&Display>)

fn get() -> DisplayManager

Gets the singleton DisplayManager object.

When called for the first time, this function consults the GDK_BACKEND environment variable to find out which of the supported GDK backends to use (in case GDK has been compiled with multiple backends). Applications can use gdk_set_allowed_backends to limit what backends can be used.

Returns

The global DisplayManager singleton; gdk_parse_args, gdk_init, or gdk_init_check must have been called first.

fn connect_display_opened<F: Fn(&DisplayManager, &Display) + 'static>(&self,
f: F)
-> u64