Struct gdk::DragContext

#[repr(transparent)]
pub struct DragContext { /* private fields */ }

Signals

action-changed

A new action is being chosen for the drag and drop operation.

This signal will only be emitted if the DragContext manages the drag and drop operation. See DragContext::manage_dnd() for more information.

cancel

The drag and drop operation was cancelled.

This signal will only be emitted if the DragContext manages the drag and drop operation. See DragContext::manage_dnd() for more information.

dnd-finished

The drag and drop operation was finished, the drag destination finished reading all data. The drag source can now free all miscellaneous data.

This signal will only be emitted if the DragContext manages the drag and drop operation. See DragContext::manage_dnd() for more information.

drop-performed

The drag and drop operation was performed on an accepting client.

This signal will only be emitted if the DragContext manages the drag and drop operation. See DragContext::manage_dnd() for more information.

Implementations

impl DragContext

pub fn actions(&self) -> DragAction

Determines the bitmask of actions proposed by the source if suggested_action() returns DragAction::ASK.

Returns

the DragAction flags

pub fn dest_window(&self) -> Window

Returns the destination window for the DND operation.

Returns

a Window

pub fn device(&self) -> Device

Returns the Device associated to the drag context.

Returns

The Device associated to self.

pub fn drag_window(&self) -> Option<Window>

Returns the window on which the drag icon should be rendered during the drag operation. Note that the window may not be available until the drag operation has begun. GDK will move the window in accordance with the ongoing drag operation. The window is owned by self and will be destroyed when the drag operation is over.

Returns

the drag window, or None

pub fn protocol(&self) -> DragProtocol

Returns the drag protocol that is used by this context.

Returns

the drag protocol

pub fn selected_action(&self) -> DragAction

Determines the action chosen by the drag destination.

Returns

a DragAction value

pub fn source_window(&self) -> Window

Returns the Window where the DND operation started.

Returns

a Window

pub fn suggested_action(&self) -> DragAction

Determines the suggested drag action of the context.

Returns

a DragAction value

pub fn list_targets(&self) -> Vec<Atom>

Retrieves the list of targets of the context.

Returns

a GList of targets

pub fn manage_dnd(&self, ipc_window: &Window, actions: DragAction) -> bool

Requests the drag and drop operation to be managed by self. When a drag and drop operation becomes managed, the DragContext will internally handle all input and source-side EventDND events as required by the windowing system.

Once the drag and drop operation is managed, the drag context will emit the following signals:

The action-changed signal whenever the final action to be performed by the drag and drop operation changes.

The drop-performed signal after the user performs the drag and drop gesture (typically by releasing the mouse button).

The dnd-finished signal after the drag and drop operation concludes (after all GdkSelection transfers happen).

The cancel signal if the drag and drop operation is finished but doesn’t happen over an accepting destination, or is cancelled through other means.

ipc_window

Window to use for IPC messaging/events

actions

the actions supported by the drag source

Returns

true if the drag and drop operation is managed.

pub fn set_device(&self, device: &impl IsA<Device>)

Associates a Device to self, so all Drag and Drop events for self are emitted as if they came from this device.

device

a Device

pub fn set_hotspot(&self, hot_x: i32, hot_y: i32)

Sets the position of the drag window that will be kept under the cursor hotspot. Initially, the hotspot is at the top left corner of the drag window.

hot_x

x coordinate of the drag window hotspot

hot_y

y coordinate of the drag window hotspot

pub fn connect_action_changed<F: Fn(&Self, DragAction) + 'static>( &self, f: F ) -> SignalHandlerId

A new action is being chosen for the drag and drop operation.

This signal will only be emitted if the DragContext manages the drag and drop operation. See manage_dnd() for more information.

action

The action currently chosen

pub fn connect_cancel<F: Fn(&Self, DragCancelReason) + 'static>( &self, f: F ) -> SignalHandlerId

The drag and drop operation was cancelled.

This signal will only be emitted if the DragContext manages the drag and drop operation. See manage_dnd() for more information.

reason

The reason the context was cancelled

pub fn connect_dnd_finished<F: Fn(&Self) + 'static>( &self, f: F ) -> SignalHandlerId

The drag and drop operation was finished, the drag destination finished reading all data. The drag source can now free all miscellaneous data.

This signal will only be emitted if the DragContext manages the drag and drop operation. See manage_dnd() for more information.

pub fn connect_drop_performed<F: Fn(&Self, i32) + 'static>( &self, f: F ) -> SignalHandlerId

The drag and drop operation was performed on an accepting client.

This signal will only be emitted if the DragContext manages the drag and drop operation. See manage_dnd() for more information.

time

the time at which the drop happened.

impl DragContext

pub fn drag_get_selection(&self) -> Atom

pub fn drag_abort(&self, time_: u32)

pub fn drop_reply(&self, accepted: bool, time_: u32)

pub fn drop(&self, time_: u32)

pub fn drag_find_window_for_screen( &self, drag_window: &Window, screen: &Screen, x_root: i32, y_root: i32 ) -> (Option<Window>, DragProtocol)

pub fn drag_motion( &self, dest_window: &Window, protocol: DragProtocol, x_root: i32, y_root: i32, suggested_action: DragAction, possible_actions: DragAction, time_: u32 ) -> bool

pub fn drop_finish(&self, success: bool, time_: u32)

pub fn drag_status(&self, action: DragAction, time_: u32)

pub fn drag_drop_succeeded(&self) -> bool

pub fn drag_begin(window: &Window, targets: &[Atom]) -> Option<DragContext>

pub fn drag_begin_for_device<P: IsA<Device>>( window: &Window, device: &P, targets: &[Atom] ) -> Option<DragContext>

pub fn drag_begin_from_point<P: IsA<Device>>( window: &Window, device: &P, targets: &[Atom], x_root: i32, y_root: i32 ) -> Option<DragContext>

pub fn drag_drop_done(&self, success: bool)