gdk

Struct EventButton

pub struct EventButton(/* private fields */);

Used for button press and button release events. The type field will be one of EventType::ButtonPress, EventType::_2buttonPress, EventType::_3buttonPress or EventType::ButtonRelease,

Double and triple-clicks result in a sequence of events being received. For double-clicks the order of events will be:

EventType::ButtonPress

EventType::ButtonRelease

EventType::ButtonPress

EventType::_2buttonPress

EventType::ButtonRelease

Note that the first click is received just like a normal button press, while the second click results in a EventType::_2buttonPress being received just after the EventType::ButtonPress.

Triple-clicks are very similar to double-clicks, except that EventType::_3buttonPress is inserted after the third click. The order of the events is:

EventType::ButtonPress

EventType::ButtonRelease

EventType::ButtonPress

EventType::_2buttonPress

EventType::ButtonRelease

EventType::ButtonPress

EventType::_3buttonPress

EventType::ButtonRelease

For a double click to occur, the second button press must occur within 1/4 of a second of the first. For a triple click to occur, the third button press must also occur within 1/2 second of the first button press.

Implementations

impl EventButton

pub fn position(&self) -> (f64, f64)

pub fn state(&self) -> ModifierType

pub fn time(&self) -> u32

pub fn button(&self) -> u32

pub fn device(&self) -> Option<Device>

pub fn axes(&self) -> Option<(f64, f64)>

pub fn root(&self) -> (f64, f64)