Struct gdk::EventScroll

pub struct EventScroll(_);

Generated from button presses for the buttons 4 to 7. Wheel mice are usually configured to generate button press events for buttons 4 and 5 when the wheel is turned.

Some GDK backends can also generate “smooth” scroll events, which can be recognized by the ScrollDirection::Smooth scroll direction. For these, the scroll deltas can be obtained with gdk_event_get_scroll_deltas.

Methods

impl EventScroll

fn get_time(&self) -> u32

fn get_position(&self) -> (f64, f64)

fn get_state(&self) -> ModifierType

fn get_device(&self) -> Option<Device>

fn get_delta(&self) -> (f64, f64)