Struct gdk::EventVisibility

pub struct EventVisibility(_);

[Deprecated since 3.12] Generated when the window visibility status has changed.

Deprecated since 3.12

Modern composited windowing systems with pervasive transparency make it impossible to track the visibility of a window reliably, so this event can not be guaranteed to provide useful information.

Methods

impl EventVisibility

pub fn get_state(&self) -> VisibilityState