Struct gdk::FrameClock

pub struct FrameClock(_, _);

A FrameClock tells the application when to update and repaint a window. This may be synced to the vertical refresh rate of the monitor, for example. Even when the frame clock uses a simple timer rather than a hardware-based vertical sync, the frame clock helps because it ensures everything paints at the same time (reducing the total number of frames). The frame clock can also automatically stop painting when it knows the frames will not be visible, or scale back animation framerates.

FrameClock is designed to be compatible with an OpenGL-based implementation or with mozRequestAnimationFrame in Firefox, for example.

A frame clock is idle until someone requests a frame with FrameClock::request_phase. At some later point that makes sense for the synchronization being implemented, the clock will process a frame and emit signals for each phase that has been requested. (See the signals of the FrameClock class for documentation of the phases. FrameClockPhase::Update and the FrameClock::update signal are most interesting for application writers, and are used to update the animations, using the frame time given by FrameClock::get_frame_time.

The frame time is reported in microseconds and generally in the same timescale as g_get_monotonic_time, however, it is not the same as g_get_monotonic_time. The frame time does not advance during the time a frame is being painted, and outside of a frame, an attempt is made so that all calls to FrameClock::get_frame_time that are called at a “similar” time get the same value. This means that if different animations are timed by looking at the difference in time between an initial value from FrameClock::get_frame_time and the value inside the FrameClock::update signal of the clock, they will stay exactly synchronized.

Feature: v3_8

Methods

impl FrameClock

fn get_refresh_info(&self, base_time: i64) -> (i64, i64)

Using the frame history stored in the frame clock, finds the last known presentation time and refresh interval, and assuming that presentation times are separated by the refresh interval, predicts a presentation time that is a multiple of the refresh interval after the last presentation time, and later than base_time.

Feature: v3_8

base_time

base time for determining a presentaton time

refresh_interval_return

a location to store the determined refresh interval, or None. A default refresh interval of 1/60th of a second will be stored if no history is present.

presentation_time_return

a location to store the next candidate presentation time after the given base time. 0 will be will be stored if no history is present.