Struct gdk::FrameTimings

pub struct FrameTimings(_);

A FrameTimings object holds timing information for a single frame of the application’s displays. To retrieve FrameTimings objects, use FrameClockExt::get_timings or FrameClockExt::get_current_timings. The information in FrameTimings is useful for precise synchronization of video with the event or audio streams, and for measuring quality metrics for the application’s display, such as latency and jitter.

Feature: v3_8

Methods

impl FrameTimings

pub fn get_complete(&self) -> bool

The timing information in a FrameTimings is filled in incrementally as the frame as drawn and passed off to the window system for processing and display to the user. The accessor functions for FrameTimings can return 0 to indicate an unavailable value for two reasons: either because the information is not yet available, or because it isn't available at all. Once FrameTimings::get_complete returns true for a frame, you can be certain that no further values will become available and be stored in the FrameTimings.

Feature: v3_8

Returns

true if all information that will be available for the frame has been filled in.

pub fn get_frame_counter(&self) -> i64

Gets the frame counter value of the FrameClock when this this frame was drawn.

Feature: v3_8

Returns

the frame counter value for this frame

pub fn get_frame_time(&self) -> i64

Returns the frame time for the frame. This is the time value that is typically used to time animations for the frame. See FrameClockExt::get_frame_time.

Returns

the frame time for the frame, in the timescale of g_get_monotonic_time

pub fn get_predicted_presentation_time(&self) -> i64

Gets the predicted time at which this frame will be displayed. Although no predicted time may be available, if one is available, it will be available while the frame is being generated, in contrast to FrameTimings::get_presentation_time, which is only available after the frame has been presented. In general, if you are simply animating, you should use FrameClockExt::get_frame_time rather than this function, but this function is useful for applications that want exact control over latency. For example, a movie player may want this information for Audio/Video synchronization.

Feature: v3_8

Returns

The predicted time at which the frame will be presented, in the timescale of g_get_monotonic_time, or 0 if no predicted presentation time is available.

pub fn get_presentation_time(&self) -> i64

Reurns the presentation time. This is the time at which the frame became visible to the user.

Feature: v3_8

Returns

the time the frame was displayed to the user, in the timescale of g_get_monotonic_time, or 0 if no presentation time is available. See FrameTimings::get_complete

pub fn get_refresh_interval(&self) -> i64

Gets the natural interval between presentation times for the display that this frame was displayed on. Frame presentation usually happens during the “vertical blanking interval”.

Feature: v3_8

Returns

the refresh interval of the display, in microseconds, or 0 if the refresh interval is not available. See FrameTimings::get_complete.