Struct gdk::Screen

pub struct Screen(_, _);

Screen objects are the GDK representation of the screen on which windows can be displayed and on which the pointer moves. X originally identified screens with physical screens, but nowadays it is more common to have a single Screen which combines several physical monitors (see Screen::get_n_monitors).

Screen is used throughout GDK and GTK+ to specify which screen the top level windows are to be displayed on. it is also used to query the screen specification and default settings such as the default visual (Screen::get_system_visual), the dimensions of the physical monitors (Screen::get_monitor_geometry), etc.

Methods

impl Screen

fn get_default() -> Option<Screen>

Gets the default screen for the default display. (See gdk_display_get_default ()).

Returns

a Screen, or None if there is no default display.

fn get_system_visual(&self) -> Visual

Get the system’s default visual for self. This is the visual for the root window of the display. The return value should not be freed.

Returns

the system visual

fn get_rgba_visual(&self) -> Option<Visual>

Gets a visual to use for creating windows with an alpha channel. The windowing system on which GTK+ is running may not support this capability, in which case None will be returned. Even if a non-None value is returned, its possible that the window’s alpha channel won’t be honored when displaying the window on the screen: in particular, for X an appropriate windowing manager and compositing manager must be running to provide appropriate display.

This functionality is not implemented in the Windows backend.

For setting an overall opacity for a top-level window, see Window::set_opacity.

Returns

a visual to use for windows with an alpha channel or None if the capability is not available.

fn is_composited(&self) -> bool

Returns whether windows with an RGBA visual can reasonably be expected to have their alpha channel drawn correctly on the screen.

On X11 this function returns whether a compositing manager is compositing self.

Returns

Whether windows with RGBA visuals can reasonably be expected to have their alpha channels drawn correctly on the screen.

fn get_root_window(&self) -> Window

Gets the root window of self.

Returns

the root window

fn get_display(&self) -> Display

Gets the display to which the self belongs.

Returns

the display to which self belongs

fn get_number(&self) -> i32

Gets the index of self among the screens in the display to which it belongs. (See Screen::get_display)

Returns

the index

fn get_width(&self) -> i32

Gets the width of self in pixels

Returns

the width of self in pixels.

fn get_height(&self) -> i32

Gets the height of self in pixels

Returns

the height of self in pixels.

fn get_width_mm(&self) -> i32

Gets the width of self in millimeters. Note that on some X servers this value will not be correct.

Returns

the width of self in millimeters.

fn get_height_mm(&self) -> i32

Returns the height of self in millimeters. Note that on some X servers this value will not be correct.

Returns

the heigth of self in millimeters.

fn make_display_name(&self) -> String

Determines the name to pass to Display::open to get a Display with this screen as the default screen.

Returns

a newly allocated string, free with g_free

fn get_n_monitors(&self) -> i32

Returns the number of monitors which self consists of.

Returns

number of monitors which self consists of

fn get_primary_monitor(&self) -> i32

Gets the primary monitor for self. The primary monitor is considered the monitor where the “main desktop” lives. While normal application windows typically allow the window manager to place the windows, specialized desktop applications such as panels should place themselves on the primary monitor.

If no primary monitor is configured by the user, the return value will be 0, defaulting to the first monitor.

Returns

An integer index for the primary monitor, or 0 if none is configured.

fn get_monitor_geometry(&self, monitor_num: i32) -> RectangleInt

Retrieves the Rectangle representing the size and position of the individual monitor within the entire screen area.

Monitor numbers start at 0. To obtain the number of monitors of self, use Screen::get_n_monitors.

Note that the size of the entire screen area can be retrieved via Screen::get_width and Screen::get_height.

monitor_num

the monitor number

dest

a Rectangle to be filled with the monitor geometry

fn get_monitor_workarea(&self, monitor_num: i32) -> RectangleInt

Retrieves the Rectangle representing the size and position of the “work area” on a monitor within the entire screen area.

The work area should be considered when positioning menus and similar popups, to avoid placing them below panels, docks or other desktop components.

Note that not all backends may have a concept of workarea. This function will return the monitor geometry if a workarea is not available, or does not apply.

Monitor numbers start at 0. To obtain the number of monitors of self, use Screen::get_n_monitors.

monitor_num

the monitor number

dest

a Rectangle to be filled with the monitor workarea

fn get_monitor_at_point(&self, x: i32, y: i32) -> i32

Returns the monitor number in which the point (x,y) is located.

x

the x coordinate in the virtual screen.

y

the y coordinate in the virtual screen.

Returns

the monitor number in which the point (x,y) lies, or a monitor close to (x,y) if the point is not in any monitor.

fn get_monitor_at_window(&self, window: &Window) -> i32

Returns the number of the monitor in which the largest area of the bounding rectangle of window resides.

window

a Window

Returns

the monitor number in which most of window is located, or if window does not intersect any monitors, a monitor, close to window.

fn get_monitor_width_mm(&self, monitor_num: i32) -> i32

Gets the width in millimeters of the specified monitor, if available.

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the width of the monitor, or -1 if not available

fn get_monitor_height_mm(&self, monitor_num: i32) -> i32

Gets the height in millimeters of the specified monitor.

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the height of the monitor, or -1 if not available

fn get_monitor_plug_name(&self, monitor_num: i32) -> Option<String>

Returns the output name of the specified monitor. Usually something like VGA, DVI, or TV, not the actual product name of the display device.

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

a newly-allocated string containing the name of the monitor, or None if the name cannot be determined

fn get_monitor_scale_factor(&self, monitor_num: i32) -> i32

Returns the internal scale factor that maps from monitor coordiantes to the actual device pixels. On traditional systems this is 1, but on very high density outputs this can be a higher value (often 2).

This can be used if you want to create pixel based data for a particula monitor, but most of the time you’re drawing to a window where it is better to use Window::get_scale_factor instead.

Since: 3.10

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the scale factor

fn get_resolution(&self) -> f64

Gets the resolution for font handling on the screen; see Screen::set_resolution for full details.

Returns

the current resolution, or -1 if no resolution has been set.

fn set_resolution(&self, dpi: f64)

Sets the resolution for font handling on the screen. This is a scale factor between points specified in a pango::FontDescription and cairo units. The default value is 96, meaning that a 10 point font will be 13 units high. (10 * 96. / 72. = 13.3).

dpi

the resolution in “dots per inch”. (Physical inches aren’t actually involved; the terminology is conventional.)

fn get_active_window(&self) -> Option<Window>

Returns the screen’s currently active window.

On X11, this is done by inspecting the _NET_ACTIVE_WINDOW property on the root window, as described in the Extended Window Manager Hints. If there is no currently currently active window, or the window manager does not support the _NET_ACTIVE_WINDOW hint, this function returns None.

On other platforms, this function may return None, depending on whether it is implementable on that platform.

The returned window should be unrefed using gobject::Object::unref when no longer needed.

Returns

the currently active window, or None.