Struct gdk::Window

pub struct Window(_);

Implements

WindowExt, WindowExtManual

Implementations

impl Window

pub fn beep(&self)

pub fn begin_draw_frame(&self, region: &Region) -> Option<DrawingContext>

This is supported on crate feature v3_22 only.

pub fn begin_move_drag(
&self,
button: i32,
root_x: i32,
root_y: i32,
timestamp: u32
)

pub fn begin_move_drag_for_device(
&self,
device: &Device,
button: i32,
root_x: i32,
root_y: i32,
timestamp: u32
)

pub fn begin_paint_rect(&self, rectangle: &Rectangle)

pub fn begin_paint_region(&self, region: &Region)

pub fn begin_resize_drag(
&self,
edge: WindowEdge,
button: i32,
root_x: i32,
root_y: i32,
timestamp: u32
)

pub fn begin_resize_drag_for_device(
&self,
edge: WindowEdge,
device: &Device,
button: i32,
root_x: i32,
root_y: i32,
timestamp: u32
)

pub fn coords_from_parent(&self, parent_x: f64, parent_y: f64) -> (f64, f64)

pub fn coords_to_parent(&self, x: f64, y: f64) -> (f64, f64)

pub fn create_gl_context(&self) -> Result<GLContext, Error>

This is supported on crate feature v3_16 only.

pub fn deiconify(&self)

pub fn destroy(&self)

pub fn destroy_notify(&self)

pub fn end_draw_frame(&self, context: &DrawingContext)

This is supported on crate feature v3_22 only.

pub fn end_paint(&self)

pub fn ensure_native(&self) -> bool

pub fn focus(&self, timestamp: u32)

pub fn freeze_toplevel_updates_libgtk_only(&self)

pub fn freeze_updates(&self)

pub fn fullscreen(&self)

pub fn fullscreen_on_monitor(&self, monitor: i32)

This is supported on crate feature v3_18 only.

pub fn geometry_changed(&self)

pub fn accepts_focus(&self) -> bool

pub fn children(&self) -> Vec<Window>

pub fn clip_region(&self) -> Option<Region>

pub fn is_composited(&self) -> bool

pub fn cursor(&self) -> Option<Cursor>

pub fn decorations(&self) -> Option<WMDecoration>

pub fn device_cursor(&self, device: &Device) -> Option<Cursor>

pub fn device_events(&self, device: &Device) -> EventMask

pub fn device_position(
&self,
device: &Device
) -> (Option<Window>, i32, i32, ModifierType)

pub fn device_position_double(
&self,
device: &Device
) -> (Option<Window>, f64, f64, ModifierType)

pub fn display(&self) -> Display

pub fn drag_protocol(&self) -> (DragProtocol, Window)

pub fn effective_parent(&self) -> Option<Window>

pub fn effective_toplevel(&self) -> Window

pub fn does_event_compression(&self) -> bool

pub fn events(&self) -> EventMask

pub fn gets_focus_on_map(&self) -> bool

pub fn frame_clock(&self) -> Option<FrameClock>

pub fn frame_extents(&self) -> Rectangle

pub fn fullscreen_mode(&self) -> FullscreenMode

pub fn geometry(&self) -> (i32, i32, i32, i32)

pub fn group(&self) -> Option<Window>

pub fn height(&self) -> i32

pub fn is_modal_hint(&self) -> bool

pub fn origin(&self) -> (i32, i32, i32)

pub fn parent(&self) -> Option<Window>

pub fn is_pass_through(&self) -> bool

This is supported on crate feature v3_18 only.

pub fn position(&self) -> (i32, i32)

pub fn root_coords(&self, x: i32, y: i32) -> (i32, i32)

pub fn root_origin(&self) -> (i32, i32)

pub fn scale_factor(&self) -> i32

pub fn screen(&self) -> Screen

pub fn source_events(&self, source: InputSource) -> EventMask

pub fn state(&self) -> WindowState

pub fn supports_multidevice(&self) -> bool

pub fn toplevel(&self) -> Window

pub fn type_hint(&self) -> WindowTypeHint

pub fn update_area(&self) -> Option<Region>

pub fn visible_region(&self) -> Option<Region>

pub fn visual(&self) -> Visual

pub fn width(&self) -> i32

pub fn window_type(&self) -> WindowType

pub fn has_native(&self) -> bool

pub fn hide(&self)

pub fn iconify(&self)

pub fn input_shape_combine_region(
&self,
shape_region: &Region,
offset_x: i32,
offset_y: i32
)

pub fn invalidate_maybe_recurse(
&self,
region: &Region,
child_func: Option<&mut dyn FnMut(&Window) -> bool>
)

pub fn invalidate_rect(
&self,
rect: Option<&Rectangle>,
invalidate_children: bool
)

pub fn invalidate_region(&self, region: &Region, invalidate_children: bool)

pub fn is_destroyed(&self) -> bool

pub fn is_input_only(&self) -> bool

pub fn is_shaped(&self) -> bool

pub fn is_viewable(&self) -> bool

pub fn is_visible(&self) -> bool

pub fn lower(&self)

pub fn mark_paint_from_clip(&self, cr: &Context)

This is supported on crate feature v3_16 only.

pub fn maximize(&self)

pub fn merge_child_input_shapes(&self)

pub fn merge_child_shapes(&self)

pub fn move_(&self, x: i32, y: i32)

pub fn move_region(&self, region: &Region, dx: i32, dy: i32)

pub fn move_resize(&self, x: i32, y: i32, width: i32, height: i32)

pub fn move_to_rect(
&self,
rect: &Rectangle,
rect_anchor: Gravity,
window_anchor: Gravity,
anchor_hints: AnchorHints,
rect_anchor_dx: i32,
rect_anchor_dy: i32
)

This is supported on crate feature v3_24 only.

pub fn peek_children(&self) -> Vec<Window>

pub fn process_updates(&self, update_children: bool)

pub fn raise(&self)

pub fn register_dnd(&self)

pub fn reparent(&self, new_parent: &Window, x: i32, y: i32)

pub fn resize(&self, width: i32, height: i32)

pub fn restack(&self, sibling: Option<&Window>, above: bool)

pub fn scroll(&self, dx: i32, dy: i32)

pub fn set_accept_focus(&self, accept_focus: bool)

pub fn set_background_rgba(&self, rgba: &RGBA)

pub fn set_child_input_shapes(&self)

pub fn set_child_shapes(&self)

pub fn set_composited(&self, composited: bool)

pub fn set_cursor(&self, cursor: Option<&Cursor>)

pub fn set_decorations(&self, decorations: WMDecoration)

pub fn set_device_cursor(&self, device: &Device, cursor: &Cursor)

pub fn set_device_events(&self, device: &Device, event_mask: EventMask)

pub fn set_event_compression(&self, event_compression: bool)

pub fn set_events(&self, event_mask: EventMask)

pub fn set_focus_on_map(&self, focus_on_map: bool)

pub fn set_fullscreen_mode(&self, mode: FullscreenMode)

pub fn set_functions(&self, functions: WMFunction)

pub fn set_geometry_hints(&self, geometry: &Geometry, geom_mask: WindowHints)

pub fn set_group(&self, leader: Option<&Window>)

pub fn set_icon_list(&self, pixbufs: &[Pixbuf])

pub fn set_icon_name(&self, name: Option<&str>)

pub fn set_keep_above(&self, setting: bool)

pub fn set_keep_below(&self, setting: bool)

pub fn set_modal_hint(&self, modal: bool)

pub fn set_opacity(&self, opacity: f64)

pub fn set_opaque_region(&self, region: Option<&Region>)

pub fn set_override_redirect(&self, override_redirect: bool)

pub fn set_pass_through(&self, pass_through: bool)

This is supported on crate feature v3_18 only.

pub fn set_role(&self, role: &str)

pub fn set_shadow_width(&self, left: i32, right: i32, top: i32, bottom: i32)

pub fn set_skip_pager_hint(&self, skips_pager: bool)

pub fn set_skip_taskbar_hint(&self, skips_taskbar: bool)

pub fn set_source_events(&self, source: InputSource, event_mask: EventMask)

pub fn set_startup_id(&self, startup_id: &str)

pub fn set_static_gravities(&self, use_static: bool) -> bool

pub fn set_support_multidevice(&self, support_multidevice: bool)

pub fn set_title(&self, title: &str)

pub fn set_transient_for(&self, parent: &Window)

pub fn set_type_hint(&self, hint: WindowTypeHint)

pub fn set_urgency_hint(&self, urgent: bool)

pub fn shape_combine_region(
&self,
shape_region: Option<&Region>,
offset_x: i32,
offset_y: i32
)

pub fn show(&self)

pub fn show_unraised(&self)

pub fn show_window_menu(&self, event: &mut Event) -> bool

pub fn stick(&self)

pub fn thaw_toplevel_updates_libgtk_only(&self)

pub fn thaw_updates(&self)

pub fn unfullscreen(&self)

pub fn unmaximize(&self)

pub fn unstick(&self)

pub fn withdraw(&self)

pub fn constrain_size(
geometry: &mut Geometry,
flags: WindowHints,
width: i32,
height: i32
) -> (i32, i32)

Constrains a desired width and height according to a set of geometry hints (such as minimum and maximum size).

geometry

a Geometry structure

flags

a mask indicating what portions of geometry are set

width

desired width of window

height

desired height of the window

new_width

location to store resulting width

new_height

location to store resulting height

pub fn process_all_updates()

Calls WindowExt::process_updates for all windows (see Window) in the application.

Deprecated since 3.22

pub fn set_debug_updates(setting: bool)

With update debugging enabled, calls to WindowExt::invalidate_region clear the invalidated region of the screen to a noticeable color, and GDK pauses for a short time before sending exposes to windows during WindowExt::process_updates. The net effect is that you can see the invalid region for each window and watch redraws as they occur. This allows you to diagnose inefficiencies in your application.

In essence, because the GDK rendering model prevents all flicker, if you are redrawing the same region 400 times you may never notice, aside from noticing a speed problem. Enabling update debugging causes GTK to flicker slowly and noticeably, so you can see exactly what’s being redrawn when, in what order.

The –gtk-debug=updates command line option passed to GTK+ programs enables this debug option at application startup time. That’s usually more useful than calling Window::set_debug_updates yourself, though you might want to use this function to enable updates sometime after application startup time.

Deprecated since 3.22

setting

true to turn on update debugging

pub fn connect_create_surface<F: Fn(&Window, i32, i32) -> Surface + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_pick_embedded_child<F: Fn(&Window, f64, f64) -> Option<Window> + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_cursor_notify<F: Fn(&Window) + 'static>(
&self,
f: F
) -> SignalHandlerId

impl Window

pub fn new(parent: Option<&Window>, attributes: &WindowAttr) -> Window

Creates a new Window using the attributes from attributes. See WindowAttr and WindowAttributesType for more details. Note: to use this on displays other than the default display, parent must be specified.

parent

a Window, or None to create the window as a child of the default root window for the default display.

attributes

attributes of the new window

attributes_mask

mask indicating which fields in attributes are valid

Returns

the new Window

pub fn create_similar_surface(
&self,
content: Content,
width: i32,
height: i32
) -> Option<Surface>

pub fn create_similar_image_surface(
&self,
format: Format,
width: i32,
height: i32,
scale: i32
) -> Option<Surface>