Trait gdk::AppLaunchContextExt

pub trait AppLaunchContextExt {
    fn set_desktop(&self, desktop: i32);

    fn set_display(&self, display: &Display);

    fn set_icon<'a, P: IsA<Icon> + 'a, Q: Into<Option<&'a P>>>(&self, icon: Q);

    fn set_icon_name<'a, P: Into<Option<&'a str>>>(&self, icon_name: P);

    fn set_screen(&self, screen: &Screen);

    fn set_timestamp(&self, timestamp: u32);

    fn get_property_display(&self) -> Option<Display>;

    fn connect_property_display_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;
}

Trait containing all AppLaunchContext methods.

Implementors

AppLaunchContext

Required Methods

fn set_desktop(&self, desktop: i32)

Sets the workspace on which applications will be launched when using this context when running under a window manager that supports multiple workspaces, as described in the Extended Window Manager Hints.

When the workspace is not specified or desktop is set to -1, it is up to the window manager to pick one, typically it will be the current workspace.

desktop

the number of a workspace, or -1

fn set_display(&self, display: &Display)

Sets the display on which applications will be launched when using this context. See also AppLaunchContextExt::set_screen.

Deprecated since 3.0

Use DisplayExt::get_app_launch_context instead

display

a Display

fn set_icon<'a, P: IsA<Icon> + 'a, Q: Into<Option<&'a P>>>(&self, icon: Q)

Sets the icon for applications that are launched with this context.

Window Managers can use this information when displaying startup notification.

See also AppLaunchContextExt::set_icon_name.

icon

a gio::Icon, or None

fn set_icon_name<'a, P: Into<Option<&'a str>>>(&self, icon_name: P)

Sets the icon for applications that are launched with this context. The icon_name will be interpreted in the same way as the Icon field in desktop files. See also AppLaunchContextExt::set_icon.

If both icon and icon_name are set, the icon_name takes priority. If neither icon or icon_name is set, the icon is taken from either the file that is passed to launched application or from the gio::AppInfo for the launched application itself.

icon_name

an icon name, or None

fn set_screen(&self, screen: &Screen)

Sets the screen on which applications will be launched when using this context. See also AppLaunchContextExt::set_display.

If both screen and display are set, the screen takes priority. If neither screen or display are set, the default screen and display are used.

screen

a Screen

fn set_timestamp(&self, timestamp: u32)

Sets the timestamp of self. The timestamp should ideally be taken from the event that triggered the launch.

Window managers can use this information to avoid moving the focus to the newly launched application when the user is busy typing in another window. This is also known as 'focus stealing prevention'.

timestamp

a timestamp

fn get_property_display(&self) -> Option<Display>

fn connect_property_display_notify<F: Fn(&Self) + 'static>(&self, f: F) -> u64