Trait gdk::CursorExt

pub trait CursorExt {
    fn get_cursor_type(&self) -> CursorType;

    fn get_display(&self) -> Display;

    fn get_image(&self) -> Option<Pixbuf>;

    fn get_surface(&self) -> (Option<Surface>, f64, f64);

    fn connect_property_cursor_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_display_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Cursor methods.

Implementors

Cursor

Required Methods

fn get_cursor_type(&self) -> CursorType

Returns the cursor type for this cursor.

Returns

a CursorType

fn get_display(&self) -> Display

Returns the display on which the Cursor is defined.

Returns

the Display associated to self

fn get_image(&self) -> Option<Pixbuf>

Returns a gdk_pixbuf::Pixbuf with the image used to display the cursor.

Note that depending on the capabilities of the windowing system and on the cursor, GDK may not be able to obtain the image data. In this case, None is returned.

Returns

a gdk_pixbuf::Pixbuf representing self, or None

fn get_surface(&self) -> (Option<Surface>, f64, f64)

Returns a cairo image surface with the image used to display the cursor.

Note that depending on the capabilities of the windowing system and on the cursor, GDK may not be able to obtain the image data. In this case, None is returned.

Feature: v3_10

x_hot

Location to store the hotspot x position, or None

y_hot

Location to store the hotspot y position, or None

Returns

a cairo::Surface representing self, or None

fn connect_property_cursor_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_display_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId