Trait gdk::DeviceToolExt

pub trait DeviceToolExt {
    fn get_hardware_id(&self) -> u64;

    fn get_serial(&self) -> u64;

    fn get_tool_type(&self) -> DeviceToolType;

    fn get_property_axes(&self) -> AxisFlags;

    fn get_property_hardware_id(&self) -> u64;

    fn get_property_serial(&self) -> u64;

    fn get_property_tool_type(&self) -> DeviceToolType;

    fn connect_property_axes_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_hardware_id_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_serial_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_tool_type_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all DeviceTool methods.

Feature: v3_22

Implementors

DeviceTool

Required Methods

fn get_hardware_id(&self) -> u64

Gets the hardware ID of this tool, or 0 if it's not known. When non-zero, the identificator is unique for the given tool model, meaning that two identical tools will share the same hardware_id, but will have different serial numbers (see DeviceToolExt::get_serial).

This is a more concrete (and device specific) method to identify a DeviceTool than DeviceToolExt::get_tool_type, as a tablet may support multiple devices with the same DeviceToolType, but having different hardware identificators.

Feature: v3_22

Returns

The hardware identificator of this tool.

fn get_serial(&self) -> u64

Gets the serial of this tool, this value can be used to identify a physical tool (eg. a tablet pen) across program executions.

Feature: v3_22

Returns

The serial ID for this tool

fn get_tool_type(&self) -> DeviceToolType

Gets the DeviceToolType of the tool.

Feature: v3_22

Returns

The physical type for this tool. This can be used to figure out what sort of pen is being used, such as an airbrush or a pencil.

fn get_property_axes(&self) -> AxisFlags

fn get_property_hardware_id(&self) -> u64

fn get_property_serial(&self) -> u64

fn get_property_tool_type(&self) -> DeviceToolType

fn connect_property_axes_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_hardware_id_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_serial_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_tool_type_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId