Trait gdk::DisplayExt

pub trait DisplayExt {
    fn beep(&self);

    fn close(&self);

    fn device_is_grabbed<P: IsA<Device>>(&self, device: &P) -> bool;

    fn flush(&self);

    fn get_app_launch_context(&self) -> Option<AppLaunchContext>;

    fn get_default_cursor_size(&self) -> u32;

    fn get_default_group(&self) -> Window;

    fn get_default_screen(&self) -> Screen;

    fn get_default_seat(&self) -> Option<Seat>;

    fn get_device_manager(&self) -> Option<DeviceManager>;

    fn get_event(&self) -> Option<Event>;

    fn get_maximal_cursor_size(&self) -> (u32, u32);

    fn get_monitor(&self, monitor_num: i32) -> Option<Monitor>;

    fn get_monitor_at_point(&self, x: i32, y: i32) -> Option<Monitor>;

    fn get_monitor_at_window(&self, window: &Window) -> Option<Monitor>;

    fn get_n_monitors(&self) -> i32;

    fn get_n_screens(&self) -> i32;

    fn get_name(&self) -> String;

    fn get_pointer(&self) -> (Screen, i32, i32, ModifierType);

    fn get_primary_monitor(&self) -> Option<Monitor>;

    fn get_screen(&self, screen_num: i32) -> Screen;

    fn get_window_at_pointer(&self) -> (Option<Window>, i32, i32);

    fn has_pending(&self) -> bool;

    fn is_closed(&self) -> bool;

    fn keyboard_ungrab(&self, time_: u32);

    fn list_devices(&self) -> Vec<Device>;

    fn list_seats(&self) -> Vec<Seat>;

    fn notify_startup_complete(&self, startup_id: &str);

    fn peek_event(&self) -> Option<Event>;

    fn pointer_is_grabbed(&self) -> bool;

    fn pointer_ungrab(&self, time_: u32);

    fn put_event(&self, event: &Event);

    fn request_selection_notification(&self, selection: &Atom) -> bool;

    fn set_double_click_distance(&self, distance: u32);

    fn set_double_click_time(&self, msec: u32);

    fn store_clipboard(
        &self, 
        clipboard_window: &Window, 
        time_: u32, 
        targets: &[&Atom]
    );

    fn supports_clipboard_persistence(&self) -> bool;

    fn supports_composite(&self) -> bool;

    fn supports_cursor_alpha(&self) -> bool;

    fn supports_cursor_color(&self) -> bool;

    fn supports_input_shapes(&self) -> bool;

    fn supports_selection_notification(&self) -> bool;

    fn supports_shapes(&self) -> bool;

    fn sync(&self);

    fn warp_pointer(&self, screen: &Screen, x: i32, y: i32);

    fn connect_closed<F: Fn(&Self, bool) + 'static>(&self, f: F) -> u64;

    fn connect_monitor_added<F: Fn(&Self, &Monitor) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_monitor_removed<F: Fn(&Self, &Monitor) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_opened<F: Fn(&Self) + 'static>(&self, f: F) -> u64;

    fn connect_seat_added<F: Fn(&Self, &Seat) + 'static>(&self, f: F) -> u64;

    fn connect_seat_removed<F: Fn(&Self, &Seat) + 'static>(&self, f: F) -> u64;
}

Trait containing all Display methods.

Implementors

Display

Required Methods

fn beep(&self)

Emits a short beep on self

fn close(&self)

Closes the connection to the windowing system for the given display, and cleans up associated resources.

fn device_is_grabbed<P: IsA<Device>>(&self, device: &P) -> bool

Returns true if there is an ongoing grab on device for self.

device

a Device

Returns

true if there is a grab in effect for device.

fn flush(&self)

Flushes any requests queued for the windowing system; this happens automatically when the main loop blocks waiting for new events, but if your application is drawing without returning control to the main loop, you may need to call this function explicitly. A common case where this function needs to be called is when an application is executing drawing commands from a thread other than the thread where the main loop is running.

This is most useful for X11. On windowing systems where requests are handled synchronously, this function will do nothing.

fn get_app_launch_context(&self) -> Option<AppLaunchContext>

Returns a AppLaunchContext suitable for launching applications on the given display.

Returns

a new AppLaunchContext for self. Free with gobject::Object::unref when done

fn get_default_cursor_size(&self) -> u32

Returns the default size to use for cursors on self.

Returns

the default cursor size.

fn get_default_group(&self) -> Window

Returns the default group leader window for all toplevel windows on self. This window is implicitly created by GDK. See WindowExt::set_group.

Returns

The default group leader window for self

fn get_default_screen(&self) -> Screen

Get the default Screen for self.

Returns

the default Screen object for self

fn get_default_seat(&self) -> Option<Seat>

Returns the default Seat for this display.

Feature: v3_20

Returns

the default seat.

fn get_device_manager(&self) -> Option<DeviceManager>

Returns the DeviceManager associated to self.

Deprecated since 3.20

Use DisplayExt::get_default_seat and Seat operations.

Returns

A DeviceManager, or None. This memory is owned by GDK and must not be freed or unreferenced.

fn get_event(&self) -> Option<Event>

Gets the next GdkEvent to be processed for self, fetching events from the windowing system if necessary.

Returns

the next GdkEvent to be processed, or None if no events are pending. The returned GdkEvent should be freed with gdk_event_free.

fn get_maximal_cursor_size(&self) -> (u32, u32)

Gets the maximal size to use for cursors on self.

width

the return location for the maximal cursor width

height

the return location for the maximal cursor height

fn get_monitor(&self, monitor_num: i32) -> Option<Monitor>

Gets a monitor associated with this display.

Feature: v3_22

monitor_num

number of the monitor

Returns

the Monitor, or None if monitor_num is not a valid monitor number

fn get_monitor_at_point(&self, x: i32, y: i32) -> Option<Monitor>

Gets the monitor in which the point (x, y) is located, or a nearby monitor if the point is not in any monitor.

Feature: v3_22

x

the x coordinate of the point

y

the y coordinate of the point

Returns

the monitor containing the point

fn get_monitor_at_window(&self, window: &Window) -> Option<Monitor>

Gets the monitor in which the largest area of window resides, or a monitor close to window if it is outside of all monitors.

Feature: v3_22

window

a Window

Returns

the monitor with the largest overlap with window

fn get_n_monitors(&self) -> i32

Gets the number of monitors that belong to self.

The returned number is valid until the next emission of the Display::monitor-added or Display::monitor-removed signal.

Feature: v3_22

Returns

the number of monitors

fn get_n_screens(&self) -> i32

Gets the number of screen managed by the self.

Deprecated since 3.10

The number of screens is always 1.

Returns

number of screens.

fn get_name(&self) -> String

Gets the name of the display.

Returns

a string representing the display name. This string is owned by GDK and should not be modified or freed.

fn get_pointer(&self) -> (Screen, i32, i32, ModifierType)

Gets the current location of the pointer and the current modifier mask for a given display.

Deprecated since 3.0

Use DeviceExt::get_position instead.

screen

location to store the screen that the cursor is on, or None.

x

location to store root window X coordinate of pointer, or None.

y

location to store root window Y coordinate of pointer, or None.

mask

location to store current modifier mask, or None

fn get_primary_monitor(&self) -> Option<Monitor>

Gets the primary monitor for the display.

The primary monitor is considered the monitor where the “main desktop” lives. While normal application windows typically allow the window manager to place the windows, specialized desktop applications such as panels should place themselves on the primary monitor.

Feature: v3_22

Returns

the primary monitor, or None if no primary monitor is configured by the user

fn get_screen(&self, screen_num: i32) -> Screen

Returns a screen object for one of the screens of the display.

Deprecated since 3.20

There is only one screen; use DisplayExt::get_default_screen to get it.

screen_num

the screen number

Returns

the Screen object

fn get_window_at_pointer(&self) -> (Option<Window>, i32, i32)

Obtains the window underneath the mouse pointer, returning the location of the pointer in that window in win_x, win_y for screen. Returns None if the window under the mouse pointer is not known to GDK (for example, belongs to another application).

Deprecated since 3.0

Use DeviceExt::get_window_at_position instead.

win_x

return location for x coordinate of the pointer location relative to the window origin, or None

win_y

return location for y coordinate of the pointer location relative & to the window origin, or None

Returns

the window under the mouse pointer, or None

fn has_pending(&self) -> bool

Returns whether the display has events that are waiting to be processed.

Returns

true if there are events ready to be processed.

fn is_closed(&self) -> bool

Finds out if the display has been closed.

Returns

true if the display is closed.

fn keyboard_ungrab(&self, time_: u32)

Release any keyboard grab

Deprecated since 3.0

Use DeviceExt::ungrab, together with DeviceExt::grab instead.

time_

a timestap (e.g GDK_CURRENT_TIME).

fn list_devices(&self) -> Vec<Device>

Returns the list of available input devices attached to self. The list is statically allocated and should not be freed.

Deprecated since 3.0

Use DeviceManagerExt::list_devices instead.

Returns

a list of Device

fn list_seats(&self) -> Vec<Seat>

Returns the list of seats known to self.

Feature: v3_20

Returns

the list of seats known to the Display

fn notify_startup_complete(&self, startup_id: &str)

Indicates to the GUI environment that the application has finished loading, using a given identifier.

GTK+ will call this function automatically for GtkWindow with custom startup-notification identifier unless gtk_window_set_auto_startup_notification is called to disable that feature.

startup_id

a startup-notification identifier, for which notification process should be completed

fn peek_event(&self) -> Option<Event>

Gets a copy of the first GdkEvent in the self’s event queue, without removing the event from the queue. (Note that this function will not get more events from the windowing system. It only checks the events that have already been moved to the GDK event queue.)

Returns

a copy of the first GdkEvent on the event queue, or None if no events are in the queue. The returned GdkEvent should be freed with gdk_event_free.

fn pointer_is_grabbed(&self) -> bool

Test if the pointer is grabbed.

Deprecated since 3.0

Use DisplayExt::device_is_grabbed instead.

Returns

true if an active X pointer grab is in effect

fn pointer_ungrab(&self, time_: u32)

Release any pointer grab.

Deprecated since 3.0

Use DeviceExt::ungrab, together with DeviceExt::grab instead.

time_

a timestap (e.g. GDK_CURRENT_TIME).

fn put_event(&self, event: &Event)

Appends a copy of the given event onto the front of the event queue for self.

event

a GdkEvent.

fn request_selection_notification(&self, selection: &Atom) -> bool

Request EventOwnerChange events for ownership changes of the selection named by the given atom.

selection

the Atom naming the selection for which ownership change notification is requested

Returns

whether EventOwnerChange events will be sent.

fn set_double_click_distance(&self, distance: u32)

Sets the double click distance (two clicks within this distance count as a double click and result in a EventType::2buttonPress event). See also DisplayExt::set_double_click_time. Applications should not set this, it is a global user-configured setting.

distance

distance in pixels

fn set_double_click_time(&self, msec: u32)

Sets the double click time (two clicks within this time interval count as a double click and result in a EventType::2buttonPress event). Applications should not set this, it is a global user-configured setting.

msec

double click time in milliseconds (thousandths of a second)

fn store_clipboard(
&self,
clipboard_window: &Window,
time_: u32,
targets: &[&Atom]
)

Issues a request to the clipboard manager to store the clipboard data. On X11, this is a special program that works according to the FreeDesktop Clipboard Specification.

clipboard_window

a Window belonging to the clipboard owner

time_

a timestamp

targets

an array of targets that should be saved, or None if all available targets should be saved.

n_targets

length of the targets array

fn supports_clipboard_persistence(&self) -> bool

Returns whether the speicifed display supports clipboard persistance; i.e. if it’s possible to store the clipboard data after an application has quit. On X11 this checks if a clipboard daemon is running.

Returns

true if the display supports clipboard persistance.

fn supports_composite(&self) -> bool

Returns true if WindowExt::set_composited can be used to redirect drawing on the window using compositing.

Currently this only works on X11 with XComposite and XDamage extensions available.

Deprecated since 3.16

Compositing is an outdated technology that only ever worked on X11.

Returns

true if windows may be composited.

fn supports_cursor_alpha(&self) -> bool

Returns true if cursors can use an 8bit alpha channel on self. Otherwise, cursors are restricted to bilevel alpha (i.e. a mask).

Returns

whether cursors can have alpha channels.

fn supports_cursor_color(&self) -> bool

Returns true if multicolored cursors are supported on self. Otherwise, cursors have only a forground and a background color.

Returns

whether cursors can have multiple colors.

fn supports_input_shapes(&self) -> bool

Returns true if gdk_window_input_shape_combine_mask can be used to modify the input shape of windows on self.

Returns

true if windows with modified input shape are supported

fn supports_selection_notification(&self) -> bool

Returns whether EventOwnerChange events will be sent when the owner of a selection changes.

Returns

whether EventOwnerChange events will be sent.

fn supports_shapes(&self) -> bool

Returns true if gdk_window_shape_combine_mask can be used to create shaped windows on self.

Returns

true if shaped windows are supported

fn sync(&self)

Flushes any requests queued for the windowing system and waits until all requests have been handled. This is often used for making sure that the display is synchronized with the current state of the program. Calling DisplayExt::sync before gdk_error_trap_pop makes sure that any errors generated from earlier requests are handled before the error trap is removed.

This is most useful for X11. On windowing systems where requests are handled synchronously, this function will do nothing.

fn warp_pointer(&self, screen: &Screen, x: i32, y: i32)

Warps the pointer of self to the point x,y on the screen screen, unless the pointer is confined to a window by a grab, in which case it will be moved as far as allowed by the grab. Warping the pointer creates events as if the user had moved the mouse instantaneously to the destination.

Note that the pointer should normally be under the control of the user. This function was added to cover some rare use cases like keyboard navigation support for the color picker in the GtkColorSelectionDialog.

Deprecated since 3.0

Use DeviceExt::warp instead.

screen

the screen of self to warp the pointer to

x

the x coordinate of the destination

y

the y coordinate of the destination

fn connect_closed<F: Fn(&Self, bool) + 'static>(&self, f: F) -> u64

The ::closed signal is emitted when the connection to the windowing system for display is closed.

is_error

true if the display was closed due to an error

fn connect_monitor_added<F: Fn(&Self, &Monitor) + 'static>(&self, f: F) -> u64

The ::monitor-added signal is emitted whenever a monitor is added.

Feature: v3_22

monitor

the monitor that was just added

fn connect_monitor_removed<F: Fn(&Self, &Monitor) + 'static>(&self, f: F) -> u64

The ::monitor-removed signal is emitted whenever a monitor is removed.

Feature: v3_22

monitor

the monitor that was just removed

fn connect_opened<F: Fn(&Self) + 'static>(&self, f: F) -> u64

The ::opened signal is emitted when the connection to the windowing system for display is opened.

fn connect_seat_added<F: Fn(&Self, &Seat) + 'static>(&self, f: F) -> u64

The ::seat-added signal is emitted whenever a new seat is made known to the windowing system.

Feature: v3_20

seat

the seat that was just added

fn connect_seat_removed<F: Fn(&Self, &Seat) + 'static>(&self, f: F) -> u64

The ::seat-removed signal is emitted whenever a seat is removed by the windowing system.

Feature: v3_20

seat

the seat that was just removed