Trait gdk::FrameClockExt

pub trait FrameClockExt: 'static {
    fn begin_updating(&self);

    fn end_updating(&self);

    fn get_current_timings(&self) -> Option<FrameTimings>;

    fn get_frame_counter(&self) -> i64;

    fn get_frame_time(&self) -> i64;

    fn get_history_start(&self) -> i64;

    fn get_timings(&self, frame_counter: i64) -> Option<FrameTimings>;

    fn request_phase(&self, phase: FrameClockPhase);

    fn connect_after_paint<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_before_paint<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_flush_events<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_layout<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn connect_paint<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;

    fn connect_resume_events<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_update<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId;
}

Trait containing all FrameClock methods.

Feature: v3_8

Implementors

FrameClock

Required methods

fn begin_updating(&self)

Starts updates for an animation. Until a matching call to FrameClockExt::end_updating is made, the frame clock will continually request a new frame with the FrameClockPhase::Update phase. This function may be called multiple times and frames will be requested until FrameClockExt::end_updating is called the same number of times.

Feature: v3_8

fn end_updating(&self)

Stops updates for an animation. See the documentation for FrameClockExt::begin_updating.

Feature: v3_8

fn get_current_timings(&self) -> Option<FrameTimings>

Gets the frame timings for the current frame.

Feature: v3_8

Returns

the FrameTimings for the frame currently being processed, or even no frame is being processed, for the previous frame. Before any frames have been procesed, returns None.

fn get_frame_counter(&self) -> i64

A FrameClock maintains a 64-bit counter that increments for each frame drawn.

Feature: v3_8

Returns

inside frame processing, the value of the frame counter for the current frame. Outside of frame processing, the frame counter for the last frame.

fn get_frame_time(&self) -> i64

Gets the time that should currently be used for animations. Inside the processing of a frame, it’s the time used to compute the animation position of everything in a frame. Outside of a frame, it's the time of the conceptual “previous frame,” which may be either the actual previous frame time, or if that’s too old, an updated time.

Feature: v3_8

Returns

a timestamp in microseconds, in the timescale of of g_get_monotonic_time.

fn get_history_start(&self) -> i64

FrameClock internally keeps a history of FrameTimings objects for recent frames that can be retrieved with FrameClockExt::get_timings. The set of stored frames is the set from the counter values given by FrameClockExt::get_history_start and FrameClockExt::get_frame_counter, inclusive.

Feature: v3_8

Returns

the frame counter value for the oldest frame that is available in the internal frame history of the FrameClock.

fn get_timings(&self, frame_counter: i64) -> Option<FrameTimings>

Retrieves a FrameTimings object holding timing information for the current frame or a recent frame. The FrameTimings object may not yet be complete: see FrameTimings::get_complete.

Feature: v3_8

frame_counter

the frame counter value identifying the frame to be received.

Returns

the FrameTimings object for the specified frame, or None if it is not available. See FrameClockExt::get_history_start.

fn request_phase(&self, phase: FrameClockPhase)

Asks the frame clock to run a particular phase. The signal corresponding the requested phase will be emitted the next time the frame clock processes. Multiple calls to FrameClockExt::request_phase will be combined together and only one frame processed. If you are displaying animated content and want to continually request the FrameClockPhase::Update phase for a period of time, you should use FrameClockExt::begin_updating instead, since this allows GTK+ to adjust system parameters to get maximally smooth animations.

Feature: v3_8

phase

the phase that is requested

fn connect_after_paint<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal ends processing of the frame. Applications should generally not handle this signal.

fn connect_before_paint<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal begins processing of the frame. Applications should generally not handle this signal.

fn connect_flush_events<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is used to flush pending motion events that are being batched up and compressed together. Applications should not handle this signal.

fn connect_layout<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is emitted as the second step of toolkit and application processing of the frame. Any work to update sizes and positions of application elements should be performed. GTK+ normally handles this internally.

fn connect_paint<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is emitted as the third step of toolkit and application processing of the frame. The frame is repainted. GDK normally handles this internally and produces expose events, which are turned into GTK+ ``GtkWidget::draw signals.

fn connect_resume_events<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is emitted after processing of the frame is finished, and is handled internally by GTK+ to resume normal event processing. Applications should not handle this signal.

fn connect_update<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

This signal is emitted as the first step of toolkit and application processing of the frame. Animations should be updated using FrameClockExt::get_frame_time. Applications can connect directly to this signal, or use gtk_widget_add_tick_callback as a more convenient interface.

Loading content...