Trait gdk::GLContextExt

pub trait GLContextExt {
    fn get_debug_enabled(&self) -> bool;

    fn get_display(&self) -> Option<Display>;

    fn get_forward_compatible(&self) -> bool;

    fn get_required_version(&self) -> (i32, i32);

    fn get_shared_context(&self) -> Option<GLContext>;

    fn get_use_es(&self) -> bool;

    fn get_version(&self) -> (i32, i32);

    fn get_window(&self) -> Option<Window>;

    fn is_legacy(&self) -> bool;

    fn make_current(&self);

    fn realize(&self) -> Result<(), Error>;

    fn set_debug_enabled(&self, enabled: bool);

    fn set_forward_compatible(&self, compatible: bool);

    fn set_required_version(&self, major: i32, minor: i32);

    fn set_use_es(&self, use_es: i32);

    fn connect_property_display_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_shared_context_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_window_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;
}

Trait containing all GLContext methods.

Feature: v3_16

Implementors

GLContext

Required Methods

fn get_debug_enabled(&self) -> bool

Retrieves the value set using GLContextExt::set_debug_enabled.

Feature: v3_16

Returns

true if debugging is enabled

fn get_display(&self) -> Option<Display>

Retrieves the Display the self is created for

Feature: v3_16

Returns

a Display or None

fn get_forward_compatible(&self) -> bool

Retrieves the value set using GLContextExt::set_forward_compatible.

Feature: v3_16

Returns

true if the context should be forward compatible

fn get_required_version(&self) -> (i32, i32)

Retrieves the major and minor version requested by calling GLContextExt::set_required_version.

Feature: v3_16

major

return location for the major version to request

minor

return location for the minor version to request

fn get_shared_context(&self) -> Option<GLContext>

Retrieves the GLContext that this self share data with.

Feature: v3_16

Returns

a GLContext or None

fn get_use_es(&self) -> bool

Checks whether the self is using an OpenGL or OpenGL ES profile.

Feature: v3_22

Returns

true if the GLContext is using an OpenGL ES profile

fn get_version(&self) -> (i32, i32)

Retrieves the OpenGL version of the self.

The self must be realized prior to calling this function.

Feature: v3_16

major

return location for the major version

minor

return location for the minor version

fn get_window(&self) -> Option<Window>

Retrieves the Window used by the self.

Feature: v3_16

Returns

a Window or None

fn is_legacy(&self) -> bool

Whether the GLContext is in legacy mode or not.

The GLContext must be realized before calling this function.

When realizing a GL context, GDK will try to use the OpenGL 3.2 core profile; this profile removes all the OpenGL API that was deprecated prior to the 3.2 version of the specification. If the realization is successful, this function will return false.

If the underlying OpenGL implementation does not support core profiles, GDK will fall back to a pre-3.2 compatibility profile, and this function will return true.

You can use the value returned by this function to decide which kind of OpenGL API to use, or whether to do extension discovery, or what kind of shader programs to load.

Feature: v3_20

Returns

true if the GL context is in legacy mode

fn make_current(&self)

Makes the self the current one.

Feature: v3_16

fn realize(&self) -> Result<(), Error>

Realizes the given GLContext.

It is safe to call this function on a realized GLContext.

Feature: v3_16

Returns

true if the context is realized

fn set_debug_enabled(&self, enabled: bool)

Sets whether the GLContext should perform extra validations and run time checking. This is useful during development, but has additional overhead.

The GLContext must not be realized or made current prior to calling this function.

Feature: v3_16

enabled

whether to enable debugging in the context

fn set_forward_compatible(&self, compatible: bool)

Sets whether the GLContext should be forward compatible.

Forward compatibile contexts must not support OpenGL functionality that has been marked as deprecated in the requested version; non-forward compatible contexts, on the other hand, must support both deprecated and non deprecated functionality.

The GLContext must not be realized or made current prior to calling this function.

Feature: v3_16

compatible

whether the context should be forward compatible

fn set_required_version(&self, major: i32, minor: i32)

Sets the major and minor version of OpenGL to request.

Setting major and minor to zero will use the default values.

The GLContext must not be realized or made current prior to calling this function.

Feature: v3_16

major

the major version to request

minor

the minor version to request

fn set_use_es(&self, use_es: i32)

Requests that GDK create a OpenGL ES context instead of an OpenGL one, if the platform and windowing system allows it.

The self must not have been realized.

By default, GDK will attempt to automatically detect whether the underlying GL implementation is OpenGL or OpenGL ES once the self is realized.

You should check the return value of GLContextExt::get_use_es after calling GLContextExt::realize to decide whether to use the OpenGL or OpenGL ES API, extensions, or shaders.

Feature: v3_22

use_es

whether the context should use OpenGL ES instead of OpenGL, or -1 to allow auto-detection

fn connect_property_display_notify<F: Fn(&Self) + 'static>(&self, f: F) -> u64

fn connect_property_shared_context_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_window_notify<F: Fn(&Self) + 'static>(&self, f: F) -> u64