Trait gdk::ScreenExt

pub trait ScreenExt {
    fn get_active_window(&self) -> Option<Window>;

    fn get_display(&self) -> Display;

    fn get_height(&self) -> i32;

    fn get_height_mm(&self) -> i32;

    fn get_monitor_at_point(&self, x: i32, y: i32) -> i32;

    fn get_monitor_at_window(&self, window: &Window) -> i32;

    fn get_monitor_geometry(&self, monitor_num: i32) -> Rectangle;

    fn get_monitor_height_mm(&self, monitor_num: i32) -> i32;

    fn get_monitor_plug_name(&self, monitor_num: i32) -> Option<String>;

    fn get_monitor_scale_factor(&self, monitor_num: i32) -> i32;

    fn get_monitor_width_mm(&self, monitor_num: i32) -> i32;

    fn get_monitor_workarea(&self, monitor_num: i32) -> Rectangle;

    fn get_n_monitors(&self) -> i32;

    fn get_number(&self) -> i32;

    fn get_primary_monitor(&self) -> i32;

    fn get_resolution(&self) -> f64;

    fn get_rgba_visual(&self) -> Option<Visual>;

    fn get_root_window(&self) -> Option<Window>;

    fn get_system_visual(&self) -> Option<Visual>;

    fn get_toplevel_windows(&self) -> Vec<Window>;

    fn get_width(&self) -> i32;

    fn get_width_mm(&self) -> i32;

    fn get_window_stack(&self) -> Vec<Window>;

    fn is_composited(&self) -> bool;

    fn list_visuals(&self) -> Vec<Visual>;

    fn make_display_name(&self) -> String;

    fn set_font_options<'a, P: Into<Option<&'a FontOptions>>>(&self, options: P);

    fn set_resolution(&self, dpi: f64);

    fn get_property_resolution(&self) -> f64;

    fn set_property_resolution(&self, resolution: f64);

    fn connect_composited_changed<F: Fn(&Self) + 'static>(&self, f: F) -> u64;

    fn connect_monitors_changed<F: Fn(&Self) + 'static>(&self, f: F) -> u64;

    fn connect_size_changed<F: Fn(&Self) + 'static>(&self, f: F) -> u64;

    fn connect_property_font_options_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;

    fn connect_property_resolution_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> u64;
}

Trait containing all Screen methods.

Implementors

Screen

Required Methods

fn get_active_window(&self) -> Option<Window>

Returns the screen’s currently active window.

On X11, this is done by inspecting the _NET_ACTIVE_WINDOW property on the root window, as described in the Extended Window Manager Hints. If there is no currently currently active window, or the window manager does not support the _NET_ACTIVE_WINDOW hint, this function returns None.

On other platforms, this function may return None, depending on whether it is implementable on that platform.

The returned window should be unrefed using gobject::Object::unref when no longer needed.

Deprecated since 3.22

Returns

the currently active window, or None.

fn get_display(&self) -> Display

Gets the display to which the self belongs.

Returns

the display to which self belongs

fn get_height(&self) -> i32

Gets the height of self in pixels. The returned size is in ”application pixels”, not in ”device pixels” (see ScreenExt::get_monitor_scale_factor).

Deprecated since 3.22

Use per-monitor information instead

Returns

the height of self in pixels.

fn get_height_mm(&self) -> i32

Returns the height of self in millimeters.

Note that this value is somewhat ill-defined when the screen has multiple monitors of different resolution. It is recommended to use the monitor dimensions instead.

Deprecated since 3.22

Use per-monitor information instead

Returns

the heigth of self in millimeters.

fn get_monitor_at_point(&self, x: i32, y: i32) -> i32

Returns the monitor number in which the point (x,y) is located.

Deprecated since 3.22

Use DisplayExt::get_monitor_at_point instead

x

the x coordinate in the virtual screen.

y

the y coordinate in the virtual screen.

Returns

the monitor number in which the point (x,y) lies, or a monitor close to (x,y) if the point is not in any monitor.

fn get_monitor_at_window(&self, window: &Window) -> i32

Returns the number of the monitor in which the largest area of the bounding rectangle of window resides.

Deprecated since 3.22

Use DisplayExt::get_monitor_at_window instead

window

a Window

Returns

the monitor number in which most of window is located, or if window does not intersect any monitors, a monitor, close to window.

fn get_monitor_geometry(&self, monitor_num: i32) -> Rectangle

Retrieves the Rectangle representing the size and position of the individual monitor within the entire screen area. The returned geometry is in ”application pixels”, not in ”device pixels” (see ScreenExt::get_monitor_scale_factor).

Monitor numbers start at 0. To obtain the number of monitors of self, use ScreenExt::get_n_monitors.

Note that the size of the entire screen area can be retrieved via ScreenExt::get_width and ScreenExt::get_height.

Deprecated since 3.22

Use MonitorExt::get_geometry instead

monitor_num

the monitor number

dest

a Rectangle to be filled with the monitor geometry

fn get_monitor_height_mm(&self, monitor_num: i32) -> i32

Gets the height in millimeters of the specified monitor.

Deprecated since 3.22

Use MonitorExt::get_height_mm instead

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the height of the monitor, or -1 if not available

fn get_monitor_plug_name(&self, monitor_num: i32) -> Option<String>

Returns the output name of the specified monitor. Usually something like VGA, DVI, or TV, not the actual product name of the display device.

Deprecated since 3.22

Use MonitorExt::get_model instead

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

a newly-allocated string containing the name of the monitor, or None if the name cannot be determined

fn get_monitor_scale_factor(&self, monitor_num: i32) -> i32

Returns the internal scale factor that maps from monitor coordinates to the actual device pixels. On traditional systems this is 1, but on very high density outputs this can be a higher value (often 2).

This can be used if you want to create pixel based data for a particular monitor, but most of the time you’re drawing to a window where it is better to use WindowExt::get_scale_factor instead.

Feature: v3_10

Deprecated since 3.22

Use MonitorExt::get_scale_factor instead

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the scale factor

fn get_monitor_width_mm(&self, monitor_num: i32) -> i32

Gets the width in millimeters of the specified monitor, if available.

Deprecated since 3.22

Use MonitorExt::get_width_mm instead

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the width of the monitor, or -1 if not available

fn get_monitor_workarea(&self, monitor_num: i32) -> Rectangle

Retrieves the Rectangle representing the size and position of the “work area” on a monitor within the entire screen area. The returned geometry is in ”application pixels”, not in ”device pixels” (see ScreenExt::get_monitor_scale_factor).

The work area should be considered when positioning menus and similar popups, to avoid placing them below panels, docks or other desktop components.

Note that not all backends may have a concept of workarea. This function will return the monitor geometry if a workarea is not available, or does not apply.

Monitor numbers start at 0. To obtain the number of monitors of self, use ScreenExt::get_n_monitors.

Deprecated since 3.22

Use MonitorExt::get_workarea instead

monitor_num

the monitor number

dest

a Rectangle to be filled with the monitor workarea

fn get_n_monitors(&self) -> i32

Returns the number of monitors which self consists of.

Deprecated since 3.22

Use DisplayExt::get_n_monitors instead

Returns

number of monitors which self consists of

fn get_number(&self) -> i32

Gets the index of self among the screens in the display to which it belongs. (See ScreenExt::get_display)

Deprecated since 3.22

Returns

the index

fn get_primary_monitor(&self) -> i32

Gets the primary monitor for self. The primary monitor is considered the monitor where the “main desktop” lives. While normal application windows typically allow the window manager to place the windows, specialized desktop applications such as panels should place themselves on the primary monitor.

If no primary monitor is configured by the user, the return value will be 0, defaulting to the first monitor.

Deprecated since 3.22

Use DisplayExt::get_primary_monitor instead

Returns

An integer index for the primary monitor, or 0 if none is configured.

fn get_resolution(&self) -> f64

Gets the resolution for font handling on the screen; see ScreenExt::set_resolution for full details.

Returns

the current resolution, or -1 if no resolution has been set.

fn get_rgba_visual(&self) -> Option<Visual>

Gets a visual to use for creating windows with an alpha channel. The windowing system on which GTK+ is running may not support this capability, in which case None will be returned. Even if a non-None value is returned, its possible that the window’s alpha channel won’t be honored when displaying the window on the screen: in particular, for X an appropriate windowing manager and compositing manager must be running to provide appropriate display.

This functionality is not implemented in the Windows backend.

For setting an overall opacity for a top-level window, see WindowExt::set_opacity.

Returns

a visual to use for windows with an alpha channel or None if the capability is not available.

fn get_root_window(&self) -> Option<Window>

Gets the root window of self.

Returns

the root window

fn get_system_visual(&self) -> Option<Visual>

Get the system’s default visual for self. This is the visual for the root window of the display. The return value should not be freed.

Returns

the system visual

fn get_toplevel_windows(&self) -> Vec<Window>

Obtains a list of all toplevel windows known to GDK on the screen self. A toplevel window is a child of the root window (see gdk_get_default_root_window).

The returned list should be freed with glib::List::free, but its elements need not be freed.

Returns

list of toplevel windows, free with glib::List::free

fn get_width(&self) -> i32

Gets the width of self in pixels. The returned size is in ”application pixels”, not in ”device pixels” (see ScreenExt::get_monitor_scale_factor).

Deprecated since 3.22

Use per-monitor information instead

Returns

the width of self in pixels.

fn get_width_mm(&self) -> i32

Gets the width of self in millimeters.

Note that this value is somewhat ill-defined when the screen has multiple monitors of different resolution. It is recommended to use the monitor dimensions instead.

Deprecated since 3.22

Use per-monitor information instead

Returns

the width of self in millimeters.

fn get_window_stack(&self) -> Vec<Window>

Returns a glib::List of GdkWindows representing the current window stack.

On X11, this is done by inspecting the _NET_CLIENT_LIST_STACKING property on the root window, as described in the Extended Window Manager Hints. If the window manager does not support the _NET_CLIENT_LIST_STACKING hint, this function returns None.

On other platforms, this function may return None, depending on whether it is implementable on that platform.

The returned list is newly allocated and owns references to the windows it contains, so it should be freed using glib::List::free and its windows unrefed using gobject::Object::unref when no longer needed.

Returns

a list of GdkWindows for the current window stack, or None.

fn is_composited(&self) -> bool

Returns whether windows with an RGBA visual can reasonably be expected to have their alpha channel drawn correctly on the screen.

On X11 this function returns whether a compositing manager is compositing self.

Returns

Whether windows with RGBA visuals can reasonably be expected to have their alpha channels drawn correctly on the screen.

fn list_visuals(&self) -> Vec<Visual>

Lists the available visuals for the specified self. A visual describes a hardware image data format. For example, a visual might support 24-bit color, or 8-bit color, and might expect pixels to be in a certain format.

Call glib::List::free on the return value when you’re finished with it.

Returns

a list of visuals; the list must be freed, but not its contents

fn make_display_name(&self) -> String

Determines the name to pass to Display::open to get a Display with this screen as the default screen.

Deprecated since 3.22

Returns

a newly allocated string, free with g_free

fn set_font_options<'a, P: Into<Option<&'a FontOptions>>>(&self, options: P)

Sets the default font options for the screen. These options will be set on any pango::Context’s newly created with gdk_pango_context_get_for_screen. Changing the default set of font options does not affect contexts that have already been created.

options

a cairo::FontOptions, or None to unset any previously set default font options.

fn set_resolution(&self, dpi: f64)

Sets the resolution for font handling on the screen. This is a scale factor between points specified in a pango::FontDescription and cairo units. The default value is 96, meaning that a 10 point font will be 13 units high. (10 * 96. / 72. = 13.3).

dpi

the resolution in “dots per inch”. (Physical inches aren’t actually involved; the terminology is conventional.)

fn get_property_resolution(&self) -> f64

fn set_property_resolution(&self, resolution: f64)

fn connect_composited_changed<F: Fn(&Self) + 'static>(&self, f: F) -> u64

The ::composited-changed signal is emitted when the composited status of the screen changes

fn connect_monitors_changed<F: Fn(&Self) + 'static>(&self, f: F) -> u64

The ::monitors-changed signal is emitted when the number, size or position of the monitors attached to the screen change.

Only for X11 and OS X for now. A future implementation for Win32 may be a possibility.

fn connect_size_changed<F: Fn(&Self) + 'static>(&self, f: F) -> u64

The ::size-changed signal is emitted when the pixel width or height of a screen changes.

fn connect_property_font_options_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64

fn connect_property_resolution_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> u64