Trait gdk::SeatExt

pub trait SeatExt {
    fn get_capabilities(&self) -> SeatCapabilities;

    fn get_display(&self) -> Option<Display>;

    fn get_keyboard(&self) -> Option<Device>;

    fn get_pointer(&self) -> Option<Device>;

    fn get_slaves(&self, capabilities: SeatCapabilities) -> Vec<Device>;

    fn ungrab(&self);

    fn connect_device_added<F: Fn(&Self, &Device) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_device_removed<F: Fn(&Self, &Device) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_tool_added<F: Fn(&Self, &DeviceTool) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_tool_removed<F: Fn(&Self, &DeviceTool) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_display_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Seat methods.

Feature: v3_20

Implementors

Seat

Required Methods

fn get_capabilities(&self) -> SeatCapabilities

Returns the capabilities this Seat currently has.

Feature: v3_20

Returns

the seat capabilities

fn get_display(&self) -> Option<Display>

Returns the Display this seat belongs to.

Returns

a Display. This object is owned by GTK+ and must not be freed.

fn get_keyboard(&self) -> Option<Device>

Returns the master device that routes keyboard events.

Feature: v3_20

Returns

a master Device with keyboard capabilities. This object is owned by GTK+ and must not be freed.

fn get_pointer(&self) -> Option<Device>

Returns the master device that routes pointer events.

Feature: v3_20

Returns

a master Device with pointer capabilities. This object is owned by GTK+ and must not be freed.

fn get_slaves(&self, capabilities: SeatCapabilities) -> Vec<Device>

Returns the slave devices that match the given capabilities.

Feature: v3_20

capabilities

capabilities to get devices for

Returns

A list of GdkDevices. The list must be freed with glib::List::free, the elements are owned by GDK and must not be freed.

fn ungrab(&self)

Releases a grab added through SeatExt::grab.

Feature: v3_20

fn connect_device_added<F: Fn(&Self, &Device) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::device-added signal is emitted when a new input device is related to this seat.

Feature: v3_20

device

the newly added Device.

fn connect_device_removed<F: Fn(&Self, &Device) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::device-removed signal is emitted when an input device is removed (e.g. unplugged).

Feature: v3_20

device

the just removed Device.

fn connect_tool_added<F: Fn(&Self, &DeviceTool) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::tool-added signal is emitted whenever a new tool is made known to the seat. The tool may later be assigned to a device (i.e. on proximity with a tablet). The device will emit the Device::tool-changed signal accordingly.

A same tool may be used by several devices.

Feature: v3_22

tool

the new DeviceTool known to the seat

fn connect_tool_removed<F: Fn(&Self, &DeviceTool) + 'static>(
&self,
f: F
) -> SignalHandlerId

This signal is emitted whenever a tool is no longer known to this seat.

Feature: v3_22

tool

the just removed DeviceTool

fn connect_property_display_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId