Crate gdk

Reexports

pub use prelude::*;

Modules

enums

functions

prelude
Traits intended for blanket imports.

Structs

AnchorHints

AppLaunchContext

Atom

AxisFlags

Color

Cursor

Device

DeviceManager

DeviceTool

Display

DisplayManager

DragAction

DragContext

DrawingContext

Error
A generic error capable of representing various error domains (types).

Event
A generic GDK event.

EventButton

EventConfigure

EventCrossing

EventDND

EventExpose

EventFocus

EventGrabBroken

EventKey

EventMask

EventMotion

EventOwnerChange

EventProperty

EventProximity

EventScroll

EventSelection

EventSetting

EventTouch

EventVisibility

EventWindowState

FrameClock

FrameClockPhase

FrameTimings

GLContext

Geometry

ModifierType

Monitor

RGBA

Rectangle

RgbaParseError

Screen

Seat

SeatCapabilities

Visual

WMDecoration

WMFunction

Window

WindowAttr

WindowHints

WindowState

Enums

AxisUse

ByteOrder

CrossingMode

CursorType

DeviceToolType

DeviceType

DragCancelReason

DragProtocol

EventType

FullscreenMode

GLError

GrabOwnership

GrabStatus

Gravity

InputMode

InputSource

ModifierIntent

NotifyType

OwnerChange

PropertyState

ScrollDirection

SettingAction

SubpixelLayout

VisibilityState

VisualType

WindowEdge

WindowType

WindowTypeHint

WindowWindowClass

Constants

ACTION_ASK

ACTION_COPY

ACTION_DEFAULT

ACTION_LINK

ACTION_MOVE

ACTION_PRIVATE

ALL_EVENTS_MASK

ANCHOR_FLIP

ANCHOR_FLIP_X

ANCHOR_FLIP_Y

ANCHOR_RESIZE

ANCHOR_RESIZE_X

ANCHOR_RESIZE_Y

ANCHOR_SLIDE

ANCHOR_SLIDE_X

ANCHOR_SLIDE_Y

ATOM_NONE

AXIS_FLAG_DISTANCE

AXIS_FLAG_PRESSURE

AXIS_FLAG_ROTATION

AXIS_FLAG_SLIDER

AXIS_FLAG_WHEEL

AXIS_FLAG_X

AXIS_FLAG_XTILT

AXIS_FLAG_Y

AXIS_FLAG_YTILT

BUTTON1_MASK

BUTTON1_MOTION_MASK

BUTTON2_MASK

BUTTON2_MOTION_MASK

BUTTON3_MASK

BUTTON3_MOTION_MASK

BUTTON4_MASK

BUTTON5_MASK

BUTTON_MOTION_MASK

BUTTON_PRESS_MASK

BUTTON_RELEASE_MASK

CONTROL_MASK

DECOR_ALL

DECOR_BORDER

DECOR_MAXIMIZE

DECOR_MENU

DECOR_MINIMIZE

DECOR_RESIZEH

DECOR_TITLE

ENTER_NOTIFY_MASK

EXPOSURE_MASK

FOCUS_CHANGE_MASK

FRAME_CLOCK_PHASE_AFTER_PAINT

FRAME_CLOCK_PHASE_BEFORE_PAINT

FRAME_CLOCK_PHASE_FLUSH_EVENTS

FRAME_CLOCK_PHASE_LAYOUT

FRAME_CLOCK_PHASE_NONE

FRAME_CLOCK_PHASE_PAINT

FRAME_CLOCK_PHASE_RESUME_EVENTS

FRAME_CLOCK_PHASE_UPDATE

FUNC_ALL

FUNC_CLOSE

FUNC_MAXIMIZE

FUNC_MINIMIZE

FUNC_MOVE

FUNC_RESIZE

HINT_ASPECT

HINT_BASE_SIZE

HINT_MAX_SIZE

HINT_MIN_SIZE

HINT_POS

HINT_RESIZE_INC

HINT_USER_POS

HINT_USER_SIZE

HINT_WIN_GRAVITY

HYPER_MASK

KEY_PRESS_MASK

KEY_RELEASE_MASK

LEAVE_NOTIFY_MASK

LOCK_MASK

META_MASK

MOD1_MASK

MOD2_MASK

MOD3_MASK

MOD4_MASK

MOD5_MASK

MODIFIER_MASK

MODIFIER_RESERVED_13_MASK

MODIFIER_RESERVED_14_MASK

MODIFIER_RESERVED_15_MASK

MODIFIER_RESERVED_16_MASK

MODIFIER_RESERVED_17_MASK

MODIFIER_RESERVED_18_MASK

MODIFIER_RESERVED_19_MASK

MODIFIER_RESERVED_20_MASK

MODIFIER_RESERVED_21_MASK

MODIFIER_RESERVED_22_MASK

MODIFIER_RESERVED_23_MASK

MODIFIER_RESERVED_24_MASK

MODIFIER_RESERVED_25_MASK

MODIFIER_RESERVED_29_MASK

POINTER_MOTION_HINT_MASK

POINTER_MOTION_MASK

PROPERTY_CHANGE_MASK

PROXIMITY_IN_MASK

PROXIMITY_OUT_MASK

RELEASE_MASK

SCROLL_MASK

SEAT_CAPABILITY_ALL

SEAT_CAPABILITY_ALL_POINTING

SEAT_CAPABILITY_KEYBOARD

SEAT_CAPABILITY_NONE

SEAT_CAPABILITY_POINTER

SEAT_CAPABILITY_TABLET_STYLUS

SEAT_CAPABILITY_TOUCH

SELECTION_CLIPBOARD

SELECTION_PRIMARY

SELECTION_SECONDARY

SELECTION_TYPE_ATOM

SELECTION_TYPE_BITMAP

SELECTION_TYPE_COLORMAP

SELECTION_TYPE_DRAWABLE

SELECTION_TYPE_INTEGER

SELECTION_TYPE_PIXMAP

SELECTION_TYPE_STRING

SELECTION_TYPE_WINDOW

SHIFT_MASK

SMOOTH_SCROLL_MASK

STRUCTURE_MASK

SUBSTRUCTURE_MASK

SUPER_MASK

TABLET_PAD_MASK

TARGET_BITMAP

TARGET_COLORMAP

TARGET_DRAWABLE

TARGET_PIXMAP

TARGET_STRING

TOUCHPAD_GESTURE_MASK

TOUCH_MASK

VISIBILITY_NOTIFY_MASK

WINDOW_STATE_ABOVE

WINDOW_STATE_BELOW

WINDOW_STATE_FOCUSED

WINDOW_STATE_FULLSCREEN

WINDOW_STATE_ICONIFIED

WINDOW_STATE_MAXIMIZED

WINDOW_STATE_STICKY

WINDOW_STATE_TILED

WINDOW_STATE_WITHDRAWN

Traits

AppLaunchContextExt

CursorExt

DeviceExt

DeviceManagerExt

DeviceToolExt

DisplayExt

DisplayManagerExt

DragContextExt

DrawingContextExt

FrameClockExt

GLContextExt

MonitorExt

ScreenExt

SeatExt

VisualExt

WindowExt

Functions

beep

error_trap_pop

error_trap_pop_ignored

error_trap_push

events_get_angle

events_get_center

events_get_distance

events_pending

flush

get_display

get_display_arg_name

get_program_class

get_show_events

init

keyboard_grab

keyboard_ungrab

keyval_convert_case

keyval_from_name

keyval_is_lower

keyval_is_upper

keyval_name

keyval_to_lower

keyval_to_unicode

keyval_to_upper

list_visuals

notify_startup_complete

notify_startup_complete_with_id

pango_context_get

pango_context_get_for_display

pango_context_get_for_screen

pixbuf_get_from_surface

pointer_grab

pointer_is_grabbed

pointer_ungrab

pre_parse_libgtk_only

property_delete

selection_convert

selection_owner_get

selection_owner_get_for_display

selection_owner_set

selection_owner_set_for_display

selection_send_notify

selection_send_notify_for_display

set_allowed_backends

set_double_click_time

set_initialized⚠
Informs this crate that GDK has been initialized and the current thread is the main one.

set_program_class

set_show_events

synthesize_window_state

test_render_sync

test_simulate_button

test_simulate_key

threads_enter

threads_init

threads_leave

unicode_to_keyval

utf8_to_string_target

Type Definitions

key