Crate gdk

Re-exports

pub use prelude::*;

Modules

enums

functions

prelude
Traits intended for blanket imports.

Structs

AppLaunchContext

AppLaunchContextClass

Atom

Color

Cursor

CursorClass

Device

DeviceClass

DeviceManager

DeviceManagerClass

Display

DisplayClass

DisplayManager

DisplayManagerClass

DragAction

DragContext

DragContextClass

Error
A generic error capable of representing various error domains (types).

Event
A generic GDK event.

EventButton

EventConfigure

EventCrossing

EventDND

EventExpose

EventFocus

EventGrabBroken

EventKey

EventMask

EventMotion

EventOwnerChange

EventProperty

EventProximity

EventScroll

EventSelection

EventSequence

EventSetting

EventTouch

EventVisibility

EventWindowState

FrameClock

FrameClockClass

FrameClockPhase

FrameTimings

GLContext

GLContextClass

GRange

Geometry

ModifierType

RGBA

Rectangle

RgbaParseError

Screen

ScreenClass

Visual

VisualClass

WMDecoration

WMFunction

Window

WindowAttr

WindowClass

WindowHints

WindowState

Enums

AxisUse

ByteOrder

ChangeData

CrossingMode

CursorType

DevicePadFeature

DeviceType

DragProtocol

EventType

FullscreenMode

GLError

GrabOwnership

GrabStatus

Gravity

InputMode

InputSource

ModifierIntent

NotifyType

OwnerChange

PropMode

PropertyState

ScrollDirection

SettingAction

VisibilityState

VisualType

WindowEdge

WindowType

WindowTypeHint

WindowWindowClass

Constants

ATOM_NONE

NONE_FRAME_CLOCK

NONE_WINDOW

SELECTION_CLIPBOARD

SELECTION_PRIMARY

SELECTION_SECONDARY

SELECTION_TYPE_ATOM

SELECTION_TYPE_BITMAP

SELECTION_TYPE_COLORMAP

SELECTION_TYPE_DRAWABLE

SELECTION_TYPE_INTEGER

SELECTION_TYPE_PIXMAP

SELECTION_TYPE_STRING

SELECTION_TYPE_WINDOW

TARGET_BITMAP

TARGET_COLORMAP

TARGET_DRAWABLE

TARGET_PIXMAP

TARGET_STRING

Traits

FrameClockExt

WindowExt

Functions

beep

error_trap_pop

error_trap_pop_ignored

error_trap_push

events_get_angle

events_get_center

events_get_distance

events_pending

flush

get_display_arg_name

get_program_class

get_show_events

init

keyval_convert_case

keyval_from_name

keyval_is_lower

keyval_is_upper

keyval_name

keyval_to_lower

keyval_to_unicode

keyval_to_upper

list_visuals

notify_startup_complete

notify_startup_complete_with_id

pango_context_get

pango_context_get_for_screen

pango_layout_get_clip_region

pango_layout_line_get_clip_region

pixbuf_get_from_surface

pre_parse_libgtk_only	Deprecated

property_change

property_delete

property_get

query_depths

selection_convert

selection_owner_get

selection_owner_get_for_display

selection_owner_set

selection_owner_set_for_display

selection_send_notify

selection_send_notify_for_display

set_allowed_backends

set_double_click_time

set_initialized⚠
Informs this crate that GDK has been initialized and the current thread is the main one.

set_program_class

set_show_events

setting_get

synthesize_window_state

test_render_sync

test_simulate_button

test_simulate_key

text_property_to_utf8_list_for_display

unicode_to_keyval

utf8_to_string_target

Type Definitions

key