Enum gdk::EventType

pub enum EventType {
    Nothing,
    Delete,
    Destroy,
    Expose,
    MotionNotify,
    ButtonPress,
    DoubleButtonPress,
    TripleButtonPress,
    ButtonRelease,
    KeyPress,
    KeyRelease,
    EnterNotify,
    LeaveNotify,
    FocusChange,
    Configure,
    Map,
    Unmap,
    PropertyNotify,
    SelectionClear,
    SelectionRequest,
    SelectionNotify,
    ProximityIn,
    ProximityOut,
    DragEnter,
    DragLeave,
    DragMotion,
    DragStatus,
    DropStart,
    DropFinished,
    ClientEvent,
    VisibilityNotify,
    Scroll,
    WindowState,
    Setting,
    OwnerChange,
    GrabBroken,
    Damage,
    TouchBegin,
    TouchUpdate,
    TouchEnd,
    TouchCancel,
    PadButtonPress,
    PadButtonRelease,
    PadRing,
    PadStrip,
    PadGroupMode,
    // some variants omitted
}

Specifies the type of the event.

Do not confuse these events with the signals that GTK+ widgets emit. Although many of these events result in corresponding signals being emitted, the events are often transformed or filtered along the way.

In some language bindings, the values EventType::2buttonPress and EventType::3buttonPress would translate into something syntactically invalid (eg Gdk.EventType.2ButtonPress, where a symbol is not allowed to start with a number). In that case, the aliases EventType::DoubleButtonPress and EventType::TripleButtonPress can be used instead.

Variants

Nothing

Delete

Destroy

Expose

MotionNotify

ButtonPress

DoubleButtonPress

TripleButtonPress

ButtonRelease

KeyPress

KeyRelease

EnterNotify

LeaveNotify

FocusChange

Configure

Map

Unmap

PropertyNotify

SelectionClear

SelectionRequest

SelectionNotify

ProximityIn

ProximityOut

DragEnter

DragLeave

DragMotion

DragStatus

DropStart

DropFinished

ClientEvent

VisibilityNotify

Scroll

WindowState

Setting

OwnerChange

GrabBroken

Damage

TouchBegin

TouchUpdate

TouchEnd

TouchCancel

PadButtonPress

PadButtonRelease

PadRing

PadStrip

PadGroupMode