Function gdk::events_get_center

pub fn events_get_center(
    event1: &mut Event, 
    event2: &mut Event
) -> Option<(f64, f64)>

If both events contain X/Y information, the center of both coordinates will be returned in x and y.

event1

first GdkEvent

event2

second GdkEvent

Returns

true if the center could be calculated.

x

return location for the X coordinate of the center

y

return location for the Y coordinate of the center