gdk
Function test_simulate_button

pub fn test_simulate_button(
    window: &Window,
    x: i32,
    y: i32,
    button: u32,
    modifiers: ModifierType,
    button_pressrelease: EventType,
) -> bool

This function is intended to be used in GTK+ test programs. It will warp the mouse pointer to the given (x,y) coordinates within window and simulate a button press or release event. Because the mouse pointer needs to be warped to the target location, use of this function outside of test programs that run in their own virtual windowing system (e.g. Xvfb) is not recommended.

Also, test_simulate_button() is a fairly low level function, for most testing purposes, gtk_test_widget_click() is the right function to call which will generate a button press event followed by its accompanying button release event.

window

a Window to simulate a button event for

x

x coordinate within window for the button event

y

y coordinate within window for the button event

button

Number of the pointer button for the event, usually 1, 2 or 3

modifiers

Keyboard modifiers the event is setup with

button_pressrelease

either EventType::ButtonPress or EventType::ButtonRelease

Returns

whether all actions necessary for a button event simulation were carried out successfully