gdk

Function test_simulate_key

pub fn test_simulate_key(
    window: &Window,
    x: i32,
    y: i32,
    keyval: u32,
    modifiers: ModifierType,
    key_pressrelease: EventType,
) -> bool

This function is intended to be used in GTK+ test programs. If (x,y) are > (-1,-1), it will warp the mouse pointer to the given (x,y) coordinates within window and simulate a key press or release event.

When the mouse pointer is warped to the target location, use of this function outside of test programs that run in their own virtual windowing system (e.g. Xvfb) is not recommended. If (x,y) are passed as (-1,-1), the mouse pointer will not be warped and window origin will be used as mouse pointer location for the event.

Also, test_simulate_key() is a fairly low level function, for most testing purposes, gtk_test_widget_send_key() is the right function to call which will generate a key press event followed by its accompanying key release event.

window

a Window to simulate a key event for

x

x coordinate within window for the key event

y

y coordinate within window for the key event

keyval

A GDK keyboard value

modifiers

Keyboard modifiers the event is setup with

key_pressrelease

either EventType::KeyPress or EventType::KeyRelease

Returns

whether all actions necessary for a key event simulation were carried out successfully