Function gdk::text_property_to_utf8_list_for_display

pub fn text_property_to_utf8_list_for_display(
    display: &Display, 
    encoding: &Atom, 
    format: i32, 
    text: &[u8]
) -> (i32, Vec<GString>)

Converts a text property in the given encoding to a list of UTF-8 strings.

display

a Display

encoding

an atom representing the encoding of the text

format

the format of the property

text

the text to convert

Returns

the number of strings in the resulting list

list

location to store the list of strings or None. The list should be freed with g_strfreev().