Trait gdk::prelude::MonitorExt

pub trait MonitorExt {
    fn get_display(&self) -> Option<Display>;

    fn get_geometry(&self) -> Rectangle;

    fn get_height_mm(&self) -> i32;

    fn get_manufacturer(&self) -> Option<String>;

    fn get_model(&self) -> Option<String>;

    fn get_refresh_rate(&self) -> i32;

    fn get_scale_factor(&self) -> i32;

    fn get_subpixel_layout(&self) -> SubpixelLayout;

    fn get_width_mm(&self) -> i32;

    fn get_workarea(&self) -> Rectangle;

    fn is_primary(&self) -> bool;

    fn get_property_display(&self) -> Option<Display>;

    fn get_property_geometry(&self) -> Option<Rectangle>;

    fn get_property_height_mm(&self) -> i32;

    fn get_property_refresh_rate(&self) -> i32;

    fn get_property_scale_factor(&self) -> i32;

    fn get_property_width_mm(&self) -> i32;

    fn get_property_workarea(&self) -> Option<Rectangle>;

    fn connect_invalidate<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_display_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_geometry_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_height_mm_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_manufacturer_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_model_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_refresh_rate_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_scale_factor_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_subpixel_layout_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_width_mm_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;

    fn connect_property_workarea_notify<F: Fn(&Self) + 'static>(
        &self, 
        f: F
    ) -> SignalHandlerId;
}

Trait containing all Monitor methods.

Feature: v3_22

Implementors

Monitor

Required Methods

fn get_display(&self) -> Option<Display>

Gets the display that this monitor belongs to.

Feature: v3_22

Returns

the display

fn get_geometry(&self) -> Rectangle

Retrieves the size and position of an individual monitor within the display coordinate space. The returned geometry is in ”application pixels”, not in ”device pixels” (see MonitorExt::get_scale_factor).

Feature: v3_22

geometry

a Rectangle to be filled wiht the monitor geometry

fn get_height_mm(&self) -> i32

Gets the height in millimeters of the monitor.

Feature: v3_22

Returns

the physical height of the monitor

fn get_manufacturer(&self) -> Option<String>

Gets the name of the monitor's manufacturer, if available.

Returns

the name of the manufacturer, or None

fn get_model(&self) -> Option<String>

Gets the a string identifying the monitor model, if available.

Returns

the monitor model, or None

fn get_refresh_rate(&self) -> i32

Gets the refresh rate of the monitor, if available.

The value is in milli-Hertz, so a refresh rate of 60Hz is returned as 60000.

Feature: v3_22

Returns

the refresh rate in milli-Hertz, or 0

fn get_scale_factor(&self) -> i32

Gets the internal scale factor that maps from monitor coordinates to the actual device pixels. On traditional systems this is 1, but on very high density outputs this can be a higher value (often 2).

This can be used if you want to create pixel based data for a particular monitor, but most of the time you’re drawing to a window where it is better to use WindowExt::get_scale_factor instead.

Feature: v3_22

Returns

the scale factor

fn get_subpixel_layout(&self) -> SubpixelLayout

Gets information about the layout of red, green and blue primaries for each pixel in this monitor, if available.

Feature: v3_22

Returns

the subpixel layout

fn get_width_mm(&self) -> i32

Gets the width in millimeters of the monitor.

Feature: v3_22

Returns

the physical width of the monitor

fn get_workarea(&self) -> Rectangle

Retrieves the size and position of the “work area” on a monitor within the display coordinate space. The returned geometry is in ”application pixels”, not in ”device pixels” (see MonitorExt::get_scale_factor).

The work area should be considered when positioning menus and similar popups, to avoid placing them below panels, docks or other desktop components.

Note that not all backends may have a concept of workarea. This function will return the monitor geometry if a workarea is not available, or does not apply.

Feature: v3_22

workarea

a Rectangle to be filled with the monitor workarea

fn is_primary(&self) -> bool

Gets whether this monitor should be considered primary (see DisplayExt::get_primary_monitor).

Feature: v3_22

Returns

true if self is primary

fn get_property_display(&self) -> Option<Display>

fn get_property_geometry(&self) -> Option<Rectangle>

fn get_property_height_mm(&self) -> i32

fn get_property_refresh_rate(&self) -> i32

fn get_property_scale_factor(&self) -> i32

fn get_property_width_mm(&self) -> i32

fn get_property_workarea(&self) -> Option<Rectangle>

fn connect_invalidate<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

fn connect_property_display_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_geometry_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_height_mm_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_manufacturer_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_model_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_refresh_rate_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_scale_factor_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_subpixel_layout_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_width_mm_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId

fn connect_property_workarea_notify<F: Fn(&Self) + 'static>(
&self,
f: F
) -> SignalHandlerId