Trait gdk::prelude::VisualExt

pub trait VisualExt {
    fn get_bits_per_rgb(&self) -> i32;

    fn get_blue_pixel_details(&self) -> (u32, i32, i32);

    fn get_byte_order(&self) -> ByteOrder;

    fn get_colormap_size(&self) -> i32;

    fn get_depth(&self) -> i32;

    fn get_green_pixel_details(&self) -> (u32, i32, i32);

    fn get_red_pixel_details(&self) -> (u32, i32, i32);

    fn get_screen(&self) -> Screen;

    fn get_visual_type(&self) -> VisualType;
}

Trait containing all Visual methods.

Implementors

Visual

Required Methods

fn get_bits_per_rgb(&self) -> i32

Deprecated

Returns the number of significant bits per red, green and blue value.

Not all GDK backend provide a meaningful value for this function.

Deprecated since 3.22

Use VisualExt::get_red_pixel_details and its variants to learn about the pixel layout of TrueColor and DirectColor visuals

Returns

The number of significant bits per color value for self.

fn get_blue_pixel_details(&self) -> (u32, i32, i32)

Obtains values that are needed to calculate blue pixel values in TrueColor and DirectColor. The “mask” is the significant bits within the pixel. The “shift” is the number of bits left we must shift a primary for it to be in position (according to the "mask"). Finally, "precision" refers to how much precision the pixel value contains for a particular primary.

mask

A pointer to a guint32 to be filled in, or None

shift

A pointer to a gint to be filled in, or None

precision

A pointer to a gint to be filled in, or None

fn get_byte_order(&self) -> ByteOrder

Deprecated

Returns the byte order of this visual.

The information returned by this function is only relevant when working with XImages, and not all backends return meaningful information for this.

Deprecated since 3.22

This information is not useful

Returns

A ByteOrder stating the byte order of self.

fn get_colormap_size(&self) -> i32

Deprecated

Returns the size of a colormap for this visual.

You have to use platform-specific APIs to manipulate colormaps.

Deprecated since 3.22

This information is not useful, since GDK does not provide APIs to operate on colormaps.

Returns

The size of a colormap that is suitable for self.

fn get_depth(&self) -> i32

Returns the bit depth of this visual.

Returns

The bit depth of this visual.

fn get_green_pixel_details(&self) -> (u32, i32, i32)

Obtains values that are needed to calculate green pixel values in TrueColor and DirectColor. The “mask” is the significant bits within the pixel. The “shift” is the number of bits left we must shift a primary for it to be in position (according to the "mask"). Finally, "precision" refers to how much precision the pixel value contains for a particular primary.

mask

A pointer to a guint32 to be filled in, or None

shift

A pointer to a gint to be filled in, or None

precision

A pointer to a gint to be filled in, or None

fn get_red_pixel_details(&self) -> (u32, i32, i32)

Obtains values that are needed to calculate red pixel values in TrueColor and DirectColor. The “mask” is the significant bits within the pixel. The “shift” is the number of bits left we must shift a primary for it to be in position (according to the "mask"). Finally, "precision" refers to how much precision the pixel value contains for a particular primary.

mask

A pointer to a guint32 to be filled in, or None

shift

A pointer to a gint to be filled in, or None

precision

A pointer to a gint to be filled in, or None

fn get_screen(&self) -> Screen

Gets the screen to which this visual belongs

Returns

the screen to which this visual belongs.

fn get_visual_type(&self) -> VisualType

Returns the type of visual this is (PseudoColor, TrueColor, etc).

Returns

A VisualType stating the type of self.