Struct gdk::Device

pub struct Device(_, _);

The Device object represents a single input device, such as a keyboard, a mouse, a touchpad, etc.

See the DeviceManager documentation for more information about the various kinds of master and slave devices, and their relationships.

Methods

impl Device

fn get_name(&self) -> Option<String>

Determines the name of the device.

Returns

a name

fn get_vendor_id(&self) -> Option<String>

Returns the vendor ID of this device, or None if this information couldn't be obtained. This ID is retrieved from the device, and is thus constant for it.

This function, together with Device::get_product_id, can be used to eg. compose gio::Settings paths to store settings for this device.

 static GSettings *
 get_device_settings (GdkDevice *device)
 {
   const gchar *vendor, *product;
   GSettings *settings;
   GdkDevice *device;
   gchar *path;

   vendor = gdk_device_get_vendor_id (device);
   product = gdk_device_get_product_id (device);

   path = g_strdup_printf ("/org/example/app/devices/%s:%s/", vendor, product);
   settings = g_settings_new_with_path (DEVICE_SCHEMA, path);
   g_free (path);

   return settings;
 }

Feature: v3_16

Returns

the vendor ID, or None

fn get_product_id(&self) -> Option<String>

Returns the product ID of this device, or None if this information couldn't be obtained. This ID is retrieved from the device, and is thus constant for it. See Device::get_vendor_id for more information.

Feature: v3_16

Returns

the product ID, or None

fn get_source(&self) -> InputSource

Determines the type of the device.

Returns

a InputSource

fn set_mode(&self, mode: InputMode)

Sets a the mode of an input device. The mode controls if the device is active and whether the device’s range is mapped to the entire screen or to a single window.

Note: This is only meaningful for floating devices, master devices (and slaves connected to these) drive the pointer cursor, which is not limited by the input mode.

mode

the input mode.

Returns

true if the mode was successfully changed.

fn get_mode(&self) -> InputMode

Determines the mode of the device.

Returns

a InputSource

fn set_key(&self, index_: u32, keyval: u32, modifiers: ModifierType)

Specifies the X key event to generate when a macro button of a device is pressed.

index_

the index of the macro button to set

keyval

the keyval to generate

modifiers

the modifiers to set

fn get_key(&self, index_: u32) -> Option<(u32, ModifierType)>

If index_ has a valid keyval, this function will return true and fill in keyval and modifiers with the keyval settings.

index_

the index of the macro button to get.

keyval

return value for the keyval.

modifiers

return value for modifiers.

Returns

true if keyval is set for index.

fn set_axis_use(&self, index_: u32, use_: AxisUse)

Specifies how an axis of a device is used.

index_

the index of the axis

use_

specifies how the axis is used

fn get_axis_use(&self, index_: u32) -> AxisUse

Returns the axis use for index_.

index_

the index of the axis.

Returns

a AxisUse specifying how the axis is used.

fn get_associated_device(&self) -> Option<Device>

Returns the associated device to self, if self is of type DeviceType::Master, it will return the paired pointer or keyboard.

If self is of type DeviceType::Slave, it will return the master device to which self is attached to.

If self is of type DeviceType::Floating, None will be returned, as there is no associated device.

Returns

The associated device, or None

fn get_device_type(&self) -> Type

Returns the device type for self.

Returns

the DeviceType for self.

fn get_display(&self) -> Display

Returns the Display to which self pertains.

Returns

a Display. This memory is owned by GTK+, and must not be freed or unreffed.

fn get_has_cursor(&self) -> bool

Determines whether the pointer follows device motion. This is not meaningful for keyboard devices, which don't have a pointer.

Returns

true if the pointer follows device motion

fn get_n_axes(&self) -> i32

Returns the number of axes the device currently has.

Returns

the number of axes.

fn get_n_keys(&self) -> i32

Returns the number of keys the device currently has.

Returns

the number of keys.

fn warp(&self, screen: &Screen, x: i32, y: i32)

Warps self in display to the point x,y on the screen screen, unless the device is confined to a window by a grab, in which case it will be moved as far as allowed by the grab. Warping the pointer creates events as if the user had moved the mouse instantaneously to the destination.

Note that the pointer should normally be under the control of the user. This function was added to cover some rare use cases like keyboard navigation support for the color picker in the GtkColorSelectionDialog.

screen

the screen to warp self to.

x

the X coordinate of the destination.

y

the Y coordinate of the destination.

fn grab(&self, window: &Window, grab_ownership: GrabOwnership, owner_events: bool, event_mask: EventMask, cursor: &Cursor, time_: u32) -> GrabStatus

Grabs the device so that all events coming from this device are passed to this application until the device is ungrabbed with Device::ungrab, or the window becomes unviewable. This overrides any previous grab on the device by this client.

Note that self and window need to be on the same display.

Device grabs are used for operations which need complete control over the given device events (either pointer or keyboard). For example in GTK+ this is used for Drag and Drop operations, popup menus and such.

Note that if the event mask of an X window has selected both button press and button release events, then a button press event will cause an automatic pointer grab until the button is released. X does this automatically since most applications expect to receive button press and release events in pairs. It is equivalent to a pointer grab on the window with owner_events set to true.

If you set up anything at the time you take the grab that needs to be cleaned up when the grab ends, you should handle the EventGrabBroken events that are emitted when the grab ends unvoluntarily.

window

the Window which will own the grab (the grab window)

grab_ownership

specifies the grab ownership.

owner_events

if false then all device events are reported with respect to window and are only reported if selected by event_mask. If true then pointer events for this application are reported as normal, but pointer events outside this application are reported with respect to window and only if selected by event_mask. In either mode, unreported events are discarded.

event_mask

specifies the event mask, which is used in accordance with owner_events.

cursor

the cursor to display while the grab is active if the device is a pointer. If this is None then the normal cursors are used for window and its descendants, and the cursor for window is used elsewhere.

time_

the timestamp of the event which led to this pointer grab. This usually comes from the GdkEvent struct, though GDK_CURRENT_TIME can be used if the time isn’t known.

Returns

GrabStatus::Success if the grab was successful.

fn ungrab(&self, time_: u32)

Release any grab on self.

time_

a timestap (e.g. GDK_CURRENT_TIME).

fn get_axis(&self, axes: &mut [f64], use_: AxisUse, value: &mut f64) -> bool

Interprets an array of double as axis values for a given device, and locates the value in the array for a given axis use.

axes

pointer to an array of axes

use_

the use to look for

value

location to store the found value.

Returns

true if the given axis use was found, otherwise false