Struct gdk::Display

pub struct Display(_, _);

Display objects purpose are two fold:

To manage and provide information about input devices (pointers and keyboards)

To manage and provide information about the available GdkScreens

Display objects are the GDK representation of an X Display, which can be described as a workstation consisting of a keyboard, a pointing device (such as a mouse) and one or more screens. It is used to open and keep track of various Screen objects currently instantiated by the application. It is also used to access the keyboard(s) and mouse pointer(s) of the display.

Most of the input device handling has been factored out into the separate DeviceManager object. Every display has a device manager, which you can obtain using DisplayExt::get_device_manager.

Implements

DisplayExt

Methods

impl Display

pub fn get_default() -> Option<Display>

Gets the default Display. This is a convenience function for: gdk_display_manager_get_default_display (gdk_display_manager_get ()).

Returns

a Display, or None if there is no default display.

pub fn open(display_name: &str) -> Option<Display>

Opens a display.

display_name

the name of the display to open

Returns

a Display, or None if the display could not be opened

pub fn open_default_libgtk_only() -> Option<Display>

Opens the default display specified by command line arguments or environment variables, sets it as the default display, and returns it. gdk_parse_args must have been called first. If the default display has previously been set, simply returns that. An internal function that should not be used by applications.

Deprecated since 3.16

This symbol was never meant to be used outside of GTK+

Returns

the default display, if it could be opened, otherwise None.