Struct gdk::Display

pub struct Display(_);

Display objects purpose are two fold:

To manage and provide information about input devices (pointers and keyboards)

To manage and provide information about the available GdkScreens

GdkDisplay objects are the GDK representation of an X Display, which can be described as a workstation consisting of a keyboard, a pointing device (such as a mouse) and one or more screens. It is used to open and keep track of various GdkScreen objects currently instantiated by the application. It is also used to access the keyboard(s) and mouse pointer(s) of the display.

Most of the input device handling has been factored out into the separate DeviceManager object. Every display has a device manager, which you can obtain using device_manager().

Implementations

impl Display

pub fn beep(&self)

Emits a short beep on self

pub fn close(&self)

Closes the connection to the windowing system for the given display, and cleans up associated resources.

pub fn device_is_grabbed(&self, device: &Device) -> bool

Returns true if there is an ongoing grab on device for self.

device

a Device

Returns

true if there is a grab in effect for device.

pub fn flush(&self)

Flushes any requests queued for the windowing system; this happens automatically when the main loop blocks waiting for new events, but if your application is drawing without returning control to the main loop, you may need to call this function explicitly. A common case where this function needs to be called is when an application is executing drawing commands from a thread other than the thread where the main loop is running.

This is most useful for X11. On windowing systems where requests are handled synchronously, this function will do nothing.

pub fn app_launch_context(&self) -> Option<AppLaunchContext>

Returns a AppLaunchContext suitable for launching applications on the given display.

Returns

a new AppLaunchContext for self. Free with g_object_unref() when done

pub fn default_cursor_size(&self) -> u32

Returns the default size to use for cursors on self.

Returns

the default cursor size.

pub fn default_group(&self) -> Window

Returns the default group leader window for all toplevel windows on self. This window is implicitly created by GDK. See Window::set_group().

Returns

The default group leader window for self

pub fn default_screen(&self) -> Screen

Get the default Screen for self.

Returns

the default Screen object for self

pub fn default_seat(&self) -> Option<Seat>

This is supported on crate feature v3_20 only.

Returns the default Seat for this display.

Returns

the default seat.

pub fn device_manager(&self) -> Option<DeviceManager>

Returns the DeviceManager associated to self.

Deprecated since 3.20

Use default_seat() and Seat operations.

Returns

A DeviceManager, or None. This memory is owned by GDK and must not be freed or unreferenced.

pub fn event(&self) -> Option<Event>

Gets the next GdkEvent to be processed for self, fetching events from the windowing system if necessary.

Returns

the next GdkEvent to be processed, or None if no events are pending. The returned GdkEvent should be freed with gdk_event_free().

pub fn maximal_cursor_size(&self) -> (u32, u32)

Gets the maximal size to use for cursors on self.

Returns

width

the return location for the maximal cursor width

height

the return location for the maximal cursor height

pub fn monitor(&self, monitor_num: i32) -> Option<Monitor>

This is supported on crate feature v3_22 only.

Gets a monitor associated with this display.

monitor_num

number of the monitor

Returns

the Monitor, or None if monitor_num is not a valid monitor number

pub fn monitor_at_point(&self, x: i32, y: i32) -> Option<Monitor>

This is supported on crate feature v3_22 only.

Gets the monitor in which the point (x, y) is located, or a nearby monitor if the point is not in any monitor.

x

the x coordinate of the point

y

the y coordinate of the point

Returns

the monitor containing the point

pub fn monitor_at_window(&self, window: &Window) -> Option<Monitor>

This is supported on crate feature v3_22 only.

Gets the monitor in which the largest area of window resides, or a monitor close to window if it is outside of all monitors.

window

a Window

Returns

the monitor with the largest overlap with window

pub fn n_monitors(&self) -> i32

This is supported on crate feature v3_22 only.

Gets the number of monitors that belong to self.

The returned number is valid until the next emission of the signal::Display::monitor-added or signal::Display::monitor-removed signal.

Returns

the number of monitors

pub fn name(&self) -> GString

Gets the name of the display.

Returns

a string representing the display name. This string is owned by GDK and should not be modified or freed.

pub fn primary_monitor(&self) -> Option<Monitor>

This is supported on crate feature v3_22 only.

Gets the primary monitor for the display.

The primary monitor is considered the monitor where the “main desktop” lives. While normal application windows typically allow the window manager to place the windows, specialized desktop applications such as panels should place themselves on the primary monitor.

Returns

the primary monitor, or None if no primary monitor is configured by the user

pub fn screen(&self, screen_num: i32) -> Screen

Returns a screen object for one of the screens of the display.

Deprecated since 3.20

There is only one screen; use default_screen() to get it.

screen_num

the screen number

Returns

the Screen object

pub fn has_pending(&self) -> bool

Returns whether the display has events that are waiting to be processed.

Returns

true if there are events ready to be processed.

pub fn is_closed(&self) -> bool

Finds out if the display has been closed.

Returns

true if the display is closed.

pub fn list_seats(&self) -> Vec<Seat>

This is supported on crate feature v3_20 only.

Returns the list of seats known to self.

Returns

the list of seats known to the Display

pub fn notify_startup_complete(&self, startup_id: &str)

Indicates to the GUI environment that the application has finished loading, using a given identifier.

GTK+ will call this function automatically for GtkWindow with custom startup-notification identifier unless gtk_window_set_auto_startup_notification() is called to disable that feature.

startup_id

a startup-notification identifier, for which notification process should be completed

pub fn peek_event(&self) -> Option<Event>

Gets a copy of the first GdkEvent in the self’s event queue, without removing the event from the queue. (Note that this function will not get more events from the windowing system. It only checks the events that have already been moved to the GDK event queue.)

Returns

a copy of the first GdkEvent on the event queue, or None if no events are in the queue. The returned GdkEvent should be freed with gdk_event_free().

pub fn put_event(&self, event: &Event)

Appends a copy of the given event onto the front of the event queue for self.

event

a GdkEvent.

pub fn request_selection_notification(&self, selection: &Atom) -> bool

Request EventOwnerChange events for ownership changes of the selection named by the given atom.

selection

the Atom naming the selection for which ownership change notification is requested

Returns

whether EventOwnerChange events will be sent.

pub fn set_double_click_distance(&self, distance: u32)

Sets the double click distance (two clicks within this distance count as a double click and result in a [EventType::_2buttonPress][crate::EventType::_2buttonPress] event). See also set_double_click_time(). Applications should not set this, it is a global user-configured setting.

distance

distance in pixels

pub fn set_double_click_time(&self, msec: u32)

Sets the double click time (two clicks within this time interval count as a double click and result in a [EventType::_2buttonPress][crate::EventType::_2buttonPress] event). Applications should not set this, it is a global user-configured setting.

msec

double click time in milliseconds (thousandths of a second)

pub fn store_clipboard(
&self,
clipboard_window: &Window,
time_: u32,
targets: &[&Atom]
)

Issues a request to the clipboard manager to store the clipboard data. On X11, this is a special program that works according to the FreeDesktop Clipboard Specification.

clipboard_window

a Window belonging to the clipboard owner

time_

a timestamp

targets

an array of targets that should be saved, or None if all available targets should be saved.

pub fn supports_clipboard_persistence(&self) -> bool

Returns whether the speicifed display supports clipboard persistance; i.e. if it’s possible to store the clipboard data after an application has quit. On X11 this checks if a clipboard daemon is running.

Returns

true if the display supports clipboard persistance.

pub fn supports_cursor_alpha(&self) -> bool

Returns true if cursors can use an 8bit alpha channel on self. Otherwise, cursors are restricted to bilevel alpha (i.e. a mask).

Returns

whether cursors can have alpha channels.

pub fn supports_cursor_color(&self) -> bool

Returns true if multicolored cursors are supported on self. Otherwise, cursors have only a forground and a background color.

Returns

whether cursors can have multiple colors.

pub fn supports_input_shapes(&self) -> bool

Returns true if gdk_window_input_shape_combine_mask() can be used to modify the input shape of windows on self.

Returns

true if windows with modified input shape are supported

pub fn supports_selection_notification(&self) -> bool

Returns whether EventOwnerChange events will be sent when the owner of a selection changes.

Returns

whether EventOwnerChange events will be sent.

pub fn supports_shapes(&self) -> bool

Returns true if gdk_window_shape_combine_mask() can be used to create shaped windows on self.

Returns

true if shaped windows are supported

pub fn sync(&self)

Flushes any requests queued for the windowing system and waits until all requests have been handled. This is often used for making sure that the display is synchronized with the current state of the program. Calling sync() before error_trap_pop() makes sure that any errors generated from earlier requests are handled before the error trap is removed.

This is most useful for X11. On windowing systems where requests are handled synchronously, this function will do nothing.

pub fn default() -> Option<Display>

Gets the default Display. This is a convenience function for: gdk_display_manager_get_default_display (gdk_display_manager_get ()).

Returns

a Display, or None if there is no default display.

pub fn open(display_name: &str) -> Option<Display>

Opens a display.

display_name

the name of the display to open

Returns

a Display, or None if the display could not be opened

pub fn connect_closed<F: Fn(&Self, bool) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::closed signal is emitted when the connection to the windowing system for display is closed.

is_error

true if the display was closed due to an error

pub fn connect_monitor_added<F: Fn(&Self, &Monitor) + 'static>(
&self,
f: F
) -> SignalHandlerId

This is supported on crate feature v3_22 only.

The ::monitor-added signal is emitted whenever a monitor is added.

monitor

the monitor that was just added

pub fn connect_monitor_removed<F: Fn(&Self, &Monitor) + 'static>(
&self,
f: F
) -> SignalHandlerId

This is supported on crate feature v3_22 only.

The ::monitor-removed signal is emitted whenever a monitor is removed.

monitor

the monitor that was just removed

pub fn connect_opened<F: Fn(&Self) + 'static>(&self, f: F) -> SignalHandlerId

The ::opened signal is emitted when the connection to the windowing system for display is opened.

pub fn connect_seat_added<F: Fn(&Self, &Seat) + 'static>(
&self,
f: F
) -> SignalHandlerId

This is supported on crate feature v3_20 only.

The ::seat-added signal is emitted whenever a new seat is made known to the windowing system.

seat

the seat that was just added

pub fn connect_seat_removed<F: Fn(&Self, &Seat) + 'static>(
&self,
f: F
) -> SignalHandlerId

This is supported on crate feature v3_20 only.

The ::seat-removed signal is emitted whenever a seat is removed by the windowing system.

seat

the seat that was just removed