Struct gdk::DragContext

pub struct DragContext(_, _);

Methods

impl DragContext

fn drag_get_selection(&self) -> Atom

fn drag_abort(&self, time_: u32)

fn drop_reply(&self, accepted: bool, time_: u32)

fn drop(&self, time_: u32)

fn drag_find_window_for_screen(&self, drag_window: &Window, screen: &Screen, x_root: i32, y_root: i32) -> (Option<Window>, DragProtocol)

fn drag_motion(&self, dest_window: &Window, protocol: DragProtocol, x_root: i32, y_root: i32, suggested_action: DragAction, possible_actions: DragAction, time_: u32) -> bool

fn drop_finish(&self, success: bool, time_: u32)

fn drag_status(&self, action: DragAction, time_: u32)

fn drag_drop_succeeded(&self) -> bool

fn get_actions(&self) -> DragAction

Determines the bitmask of actions proposed by the source if DragContext::get_suggested_action returns DragAction::Ask.

Returns

the DragAction flags

fn get_suggested_action(&self) -> DragAction

Determines the suggested drag action of the context.

Returns

a DragAction value

fn get_selected_action(&self) -> DragAction

Determines the action chosen by the drag destination.

Returns

a DragAction value

fn get_device(&self) -> Device

Returns the Device associated to the drag context.

Returns

The Device associated to self.

fn set_device(&self, device: &Device)

Associates a Device to self, so all Drag and Drop events for self are emitted as if they came from this device.

device

a Device

fn get_source_window(&self) -> Window

Returns the Window where the DND operation started.

Returns

a Window

fn get_dest_window(&self) -> Window

Returns the destination windw for the DND operation.

Returns

a Window

fn get_protocol(&self) -> DragProtocol

Returns the drag protocol thats used by this context.

Returns

the drag protocol