Struct gdk::FrameClock

pub struct FrameClock(_, _);

A FrameClock tells the application when to update and repaint a window. This may be synced to the vertical refresh rate of the monitor, for example. Even when the frame clock uses a simple timer rather than a hardware-based vertical sync, the frame clock helps because it ensures everything paints at the same time (reducing the total number of frames). The frame clock can also automatically stop painting when it knows the frames will not be visible, or scale back animation framerates.

FrameClock is designed to be compatible with an OpenGL-based implementation or with mozRequestAnimationFrame in Firefox, for example.

A frame clock is idle until someone requests a frame with FrameClock::request_phase. At some later point that makes sense for the synchronization being implemented, the clock will process a frame and emit signals for each phase that has been requested. (See the signals of the FrameClock class for documentation of the phases. FrameClockPhase::Update and the FrameClock::update signal are most interesting for application writers, and are used to update the animations, using the frame time given by FrameClock::get_frame_time.

The frame time is reported in microseconds and generally in the same timescale as g_get_monotonic_time, however, it is not the same as g_get_monotonic_time. The frame time does not advance during the time a frame is being painted, and outside of a frame, an attempt is made so that all calls to FrameClock::get_frame_time that are called at a “similar” time get the same value. This means that if different animations are timed by looking at the difference in time between an initial value from FrameClock::get_frame_time and the value inside the FrameClock::update signal of the clock, they will stay exactly synchronized.

Feature: v3_8

Methods

impl FrameClock

fn begin_updating(&self)

Starts updates for an animation. Until a matching call to FrameClock::end_updating is made, the frame clock will continually request a new frame with the FrameClockPhase::Update phase. This function may be called multiple times and frames will be requested until FrameClock::end_updating is called the same number of times.

Feature: v3_8

fn end_updating(&self)

Stops updates for an animation. See the documentation for FrameClock::begin_updating.

Feature: v3_8

fn get_current_timings(&self) -> Option<FrameTimings>

Gets the frame timings for the current frame.

Feature: v3_8

Returns

the FrameTimings for the frame currently being processed, or even no frame is being processed, for the previous frame. Before any frames have been procesed, returns None.

fn get_frame_counter(&self) -> i64

A FrameClock maintains a 64-bit counter that increments for each frame drawn.

Feature: v3_8

Returns

inside frame processing, the value of the frame counter for the current frame. Outside of frame processing, the frame counter for the last frame.

fn get_frame_time(&self) -> i64

Gets the time that should currently be used for animations. Inside the processing of a frame, it’s the time used to compute the animation position of everything in a frame. Outside of a frame, it's the time of the conceptual “previous frame,” which may be either the actual previous frame time, or if that’s too old, an updated time.

Feature: v3_8

Returns

a timestamp in microseconds, in the timescale of of g_get_monotonic_time.

fn get_history_start(&self) -> i64

FrameClock internally keeps a history of FrameTimings objects for recent frames that can be retrieved with FrameClock::get_timings. The set of stored frames is the set from the counter values given by FrameClock::get_history_start and FrameClock::get_frame_counter, inclusive.

Feature: v3_8

Returns

the frame counter value for the oldest frame that is available in the internal frame history of the FrameClock.

fn get_timings(&self, frame_counter: i64) -> Option<FrameTimings>

Retrieves a FrameTimings object holding timing information for the current frame or a recent frame. The FrameTimings object may not yet be complete: see FrameTimings::get_complete.

Feature: v3_8

frame_counter

the frame counter value identifying the frame to be received.

Returns

the FrameTimings object for the specified frame, or None if it is not available. See FrameClock::get_history_start.

fn request_phase(&self, phase: FrameClockPhase)

Asks the frame clock to run a particular phase. The signal corresponding the requested phase will be emitted the next time the frame clock processes. Multiple calls to FrameClock::request_phase will be combined together and only one frame processed. If you are displaying animated content and want to continually request the FrameClockPhase::Update phase for a period of time, you should use FrameClock::begin_updating instead, since this allows GTK+ to adjust system parameters to get maximally smooth animations.

Feature: v3_8

phase

the phase that is requested

fn connect_after_paint<F: Fn(&FrameClock) + 'static>(&self, f: F) -> u64

fn connect_before_paint<F: Fn(&FrameClock) + 'static>(&self, f: F) -> u64

fn connect_flush_events<F: Fn(&FrameClock) + 'static>(&self, f: F) -> u64

fn connect_layout<F: Fn(&FrameClock) + 'static>(&self, f: F) -> u64

fn connect_paint<F: Fn(&FrameClock) + 'static>(&self, f: F) -> u64

fn connect_resume_events<F: Fn(&FrameClock) + 'static>(&self, f: F) -> u64

fn connect_update<F: Fn(&FrameClock) + 'static>(&self, f: F) -> u64

impl FrameClock

fn get_refresh_info(&self, base_time: i64) -> (i64, i64)

Using the frame history stored in the frame clock, finds the last known presentation time and refresh interval, and assuming that presentation times are separated by the refresh interval, predicts a presentation time that is a multiple of the refresh interval after the last presentation time, and later than base_time.

Feature: v3_8

base_time

base time for determining a presentaton time

refresh_interval_return

a location to store the determined refresh interval, or None. A default refresh interval of 1/60th of a second will be stored if no history is present.

presentation_time_return

a location to store the next candidate presentation time after the given base time. 0 will be will be stored if no history is present.