Struct gdk::FrameClock

pub struct FrameClock(_, _);

A FrameClock tells the application when to update and repaint a window. This may be synced to the vertical refresh rate of the monitor, for example. Even when the frame clock uses a simple timer rather than a hardware-based vertical sync, the frame clock helps because it ensures everything paints at the same time (reducing the total number of frames). The frame clock can also automatically stop painting when it knows the frames will not be visible, or scale back animation framerates.

FrameClock is designed to be compatible with an OpenGL-based implementation or with mozRequestAnimationFrame in Firefox, for example.

A frame clock is idle until someone requests a frame with FrameClockExt::request_phase. At some later point that makes sense for the synchronization being implemented, the clock will process a frame and emit signals for each phase that has been requested. (See the signals of the FrameClock class for documentation of the phases. FrameClockPhase::Update and the FrameClock::update signal are most interesting for application writers, and are used to update the animations, using the frame time given by FrameClockExt::get_frame_time.

The frame time is reported in microseconds and generally in the same timescale as g_get_monotonic_time, however, it is not the same as g_get_monotonic_time. The frame time does not advance during the time a frame is being painted, and outside of a frame, an attempt is made so that all calls to FrameClockExt::get_frame_time that are called at a “similar” time get the same value. This means that if different animations are timed by looking at the difference in time between an initial value from FrameClockExt::get_frame_time and the value inside the FrameClock::update signal of the clock, they will stay exactly synchronized.

Feature: v3_8

Implements

FrameClockExt

Methods

impl FrameClock

pub fn get_refresh_info(&self, base_time: i64) -> (i64, i64)