Struct gdk::GLContext

pub struct GLContext(_, _);

GLContext is an object representing the platform-specific OpenGL drawing context.

GdkGLContexts are created for a Window using Window::create_gl_context, and the context will match the Visual of the window.

A GLContext is not tied to any particular normal framebuffer. For instance, it cannot draw to the Window back buffer. The GDK repaint system is in full control of the painting to that. Instead, you can create render buffers or textures and use gdk_cairo_draw_from_gl in the draw function of your widget to draw them. Then GDK will handle the integration of your rendering with that of other widgets.

Support for GLContext is platform-specific, context creation can fail, returning None context.

A GLContext has to be made "current" in order to start using it, otherwise any OpenGL call will be ignored.

Creating a new OpenGL context

In order to create a new GLContext instance you need a Window, which you typically get during the realize call of a widget.

A GLContext is not realized until either GLContext::make_current, or until it is realized using GLContext::realize. It is possible to specify details of the GL context like the OpenGL version to be used, or whether the GL context should have extra state validation enabled after calling Window::create_gl_context by calling GLContext::realize. If the realization fails you have the option to change the settings of the GLContext and try again.

Using a GLContext

You will need to make the GLContext the current context before issuing OpenGL calls; the system sends OpenGL commands to whichever context is current. It is possible to have multiple contexts, so you always need to ensure that the one which you want to draw with is the current one before issuing commands:

  gdk_gl_context_make_current (context);

You can now perform your drawing using OpenGL commands.

You can check which GLContext is the current one by using GLContext::get_current; you can also unset any GLContext that is currently set by calling GLContext::clear_current.

Since: 3.16

Methods

impl GLContext

fn get_display(&self) -> Option<Display>

Retrieves the Display the self is created for

Since: 3.16

Returns

a Display or None

fn get_window(&self) -> Option<Window>

Retrieves the Window used by the self.

Since: 3.16

Returns

a Window or None

fn get_shared_context(&self) -> Option<GLContext>

Retrieves the GLContext that this self share data with.

Since: 3.16

Returns

a GLContext or None

fn get_version(&self) -> (i32, i32)

Retrieves the OpenGL version of the self.

The self must be realized prior to calling this function.

Since: 3.16

major

return location for the major version

minor

return location for the minor version

fn set_required_version(&self, major: i32, minor: i32)

Sets the major and minor version of OpenGL to request.

Setting major and minor to zero will use the default values.

The GLContext must not be realized or made current prior to calling this function.

Since: 3.16

major

the major version to request

minor

the minor version to request

fn set_debug_enabled(&self, enabled: bool)

Sets whether the GLContext should perform extra validations and run time checking. This is useful during development, but has additional overhead.

The GLContext must not be realized or made current prior to calling this function.

Since: 3.16

enabled

whether to enable debugging in the context

fn get_debug_enabled(&self) -> bool

Retrieves the value set using GLContext::set_debug_enabled.

Since: 3.16

Returns

true if debugging is enabled

fn set_forward_compatible(&self, compatible: bool)

Sets whether the GLContext should be forward compatible.

Forward compatibile contexts must not support OpenGL functionality that has been marked as deprecated in the requested version; non-forward compatible contexts, on the other hand, must support both deprecated and non deprecated functionality.

The GLContext must not be realized or made current prior to calling this function.

Since: 3.16

compatible

whether the context should be forward compatible

fn get_forward_compatible(&self) -> bool

Retrieves the value set using GLContext::set_forward_compatible.

Since: 3.16

Returns

true if the context should be forward compatible

fn make_current(&self)

Makes the self the current one.

Since: 3.16

fn get_current() -> Option<GLContext>

Retrieves the current GLContext.

Since: 3.16

Returns

the current GLContext, or None

fn clear_current()

Clears the current GLContext.

Any OpenGL call after this function returns will be ignored until GLContext::make_current is called.

Since: 3.16