Struct gdk::RGBA

#[repr(transparent)]
pub struct RGBA { /* private fields */ }

A RGBA is used to represent a (possibly translucent) color, in a way that is compatible with cairo’s notion of color.

Implementations

impl RGBA

pub fn to_str(&self) -> GString

Returns a textual specification of self in the form rgb(r,g,b) or rgba(r g,b,a), where “r”, “g”, “b” and “a” represent the red, green, blue and alpha values respectively. “r”, “g”, and “b” are represented as integers in the range 0 to 255, and “a” is represented as a floating point value in the range 0 to 1.

These string forms are string forms that are supported by the CSS3 colors module, and can be parsed by parse().

Note that this string representation may lose some precision, since “r”, “g” and “b” are represented as 8-bit integers. If this is a concern, you should use a different representation.

Returns

A newly allocated text string

impl RGBA

pub fn new(red: f64, green: f64, blue: f64, alpha: f64) -> RGBA

pub fn red(&self) -> f64

pub fn set_red(&mut self, red: f64)

pub fn green(&self) -> f64

pub fn set_green(&mut self, green: f64)

pub fn blue(&self) -> f64

pub fn set_blue(&mut self, blue: f64)

pub fn alpha(&self) -> f64

pub fn set_alpha(&mut self, alpha: f64)

pub fn parse(s: &str) -> Result<Self, BoolError>

Parses a textual representation of a color, filling in the red, green, blue and alpha fields of the self RGBA.

The string can be either one of:

A standard name (Taken from the X11 rgb.txt file).

A hexadecimal value in the form “`rgb”, “\rrggbb”, “\rrrgggbbb” or ”\rrrrggggbbbb`”

A RGB color in the form “rgb(r,g,b)” (In this case the color will have full opacity)

A RGBA color in the form “rgba(r,g,b,a)”

Where “r”, “g”, “b” and “a” are respectively the red, green, blue and alpha color values. In the last two cases, “r”, “g”, and “b” are either integers in the range 0 to 255 or percentage values in the range 0% to 100%, and a is a floating point value in the range 0 to 1.

spec

the string specifying the color

Returns

true if the parsing succeeded

pub const BLACK: RGBA = RGBA { inner: ffi::GdkRGBA { red: 0f64, green: 0f64, blue: 0f64, alpha: 1f64, }, phantom: std::marker::PhantomData, }

pub const BLUE: RGBA = RGBA { inner: ffi::GdkRGBA { red: 0f64, green: 0f64, blue: 1f64, alpha: 1f64, }, phantom: std::marker::PhantomData, }

pub const GREEN: RGBA = RGBA { inner: ffi::GdkRGBA { red: 0f64, green: 1f64, blue: 0f64, alpha: 1f64, }, phantom: std::marker::PhantomData, }

pub const RED: RGBA = RGBA { inner: ffi::GdkRGBA { red: 1f64, green: 0f64, blue: 0f64, alpha: 1f64, }, phantom: std::marker::PhantomData, }

pub const WHITE: RGBA = RGBA { inner: ffi::GdkRGBA { red: 1f64, green: 1f64, blue: 1f64, alpha: 1f64, }, phantom: std::marker::PhantomData, }