Struct gdk::Screen

pub struct Screen(_, _);

Screen objects are the GDK representation of the screen on which windows can be displayed and on which the pointer moves. X originally identified screens with physical screens, but nowadays it is more common to have a single Screen which combines several physical monitors (see ScreenExt::get_n_monitors).

Screen is used throughout GDK and GTK+ to specify which screen the top level windows are to be displayed on. it is also used to query the screen specification and default settings such as the default visual (ScreenExt::get_system_visual), the dimensions of the physical monitors (ScreenExt::get_monitor_geometry), etc.

Implements

ScreenExt

Methods

impl Screen

pub fn get_active_window(&self) -> Option<Window>

pub fn get_display(&self) -> Display

pub fn get_height(&self) -> i32

pub fn get_height_mm(&self) -> i32

pub fn get_monitor_at_point(&self, x: i32, y: i32) -> i32

pub fn get_monitor_at_window<P: IsA<Window>>(&self, window: &P) -> i32

pub fn get_monitor_geometry(&self, monitor_num: i32) -> Rectangle

pub fn get_monitor_height_mm(&self, monitor_num: i32) -> i32

pub fn get_monitor_plug_name(&self, monitor_num: i32) -> Option<GString>

pub fn get_monitor_scale_factor(&self, monitor_num: i32) -> i32

pub fn get_monitor_width_mm(&self, monitor_num: i32) -> i32

pub fn get_monitor_workarea(&self, monitor_num: i32) -> Rectangle

pub fn get_n_monitors(&self) -> i32

pub fn get_number(&self) -> i32

pub fn get_primary_monitor(&self) -> i32

pub fn get_resolution(&self) -> f64

pub fn get_rgba_visual(&self) -> Option<Visual>

pub fn get_root_window(&self) -> Option<Window>

pub fn get_system_visual(&self) -> Option<Visual>

pub fn get_toplevel_windows(&self) -> Vec<Window>

pub fn get_width(&self) -> i32

pub fn get_width_mm(&self) -> i32

pub fn get_window_stack(&self) -> Vec<Window>

pub fn is_composited(&self) -> bool

pub fn list_visuals(&self) -> Vec<Visual>

pub fn make_display_name(&self) -> GString

pub fn set_font_options<'a, P: Into<Option<&'a FontOptions>>>(&self, options: P)

pub fn set_resolution(&self, dpi: f64)

pub fn get_default() -> Option<Screen>

Gets the default screen for the default display. (See gdk_display_get_default ()).

Returns

a Screen, or None if there is no default display.

pub fn height() -> i32

Gets the height of the default screen in pixels. The returned size is in ”application pixels”, not in ”device pixels” (see ScreenExt::get_monitor_scale_factor).

Deprecated since 3.22

Use per-monitor information

Returns

the height of the default screen in pixels.

pub fn height_mm() -> i32

Returns the height of the default screen in millimeters. Note that on many X servers this value will not be correct.

Deprecated since 3.22

Use per-monitor information

Returns

the height of the default screen in millimeters, though it is not always correct.

pub fn width() -> i32

Gets the width of the default screen in pixels. The returned size is in ”application pixels”, not in ”device pixels” (see ScreenExt::get_monitor_scale_factor).

Deprecated since 3.22

Use per-monitor information

Returns

the width of the default screen in pixels.

pub fn width_mm() -> i32

Returns the width of the default screen in millimeters. Note that on many X servers this value will not be correct.

Deprecated since 3.22

Use per-monitor information

Returns

the width of the default screen in millimeters, though it is not always correct.

pub fn connect_composited_changed<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_monitors_changed<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_size_changed<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_font_options_notify<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_resolution_notify<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

impl Screen

pub fn get_font_options(&self) -> Option<FontOptions>

pub fn get_setting(&self, name: &str) -> Option<Value>