Struct gdk::Screen

pub struct Screen(_);

Implementations

impl Screen

pub fn get_active_window(&self) -> Option<Window>

👎 Deprecated

Returns the screen’s currently active window.

On X11, this is done by inspecting the _NET_ACTIVE_WINDOW property on the root window, as described in the Extended Window Manager Hints. If there is no currently currently active window, or the window manager does not support the _NET_ACTIVE_WINDOW hint, this function returns None.

On other platforms, this function may return None, depending on whether it is implementable on that platform.

The returned window should be unrefed using gobject::Object::unref when no longer needed.

Deprecated since 3.22

Returns

the currently active window, or None.

pub fn get_display(&self) -> Display

Gets the display to which the self belongs.

Returns

the display to which self belongs

pub fn get_height(&self) -> i32

👎 Deprecated

Gets the height of self in pixels. The returned size is in ”application pixels”, not in ”device pixels” (see Screen::get_monitor_scale_factor).

Deprecated since 3.22

Use per-monitor information instead

Returns

the height of self in pixels.

pub fn get_height_mm(&self) -> i32

👎 Deprecated

Returns the height of self in millimeters.

Note that this value is somewhat ill-defined when the screen has multiple monitors of different resolution. It is recommended to use the monitor dimensions instead.

Deprecated since 3.22

Use per-monitor information instead

Returns

the heigth of self in millimeters.

pub fn get_monitor_at_point(&self, x: i32, y: i32) -> i32

👎 Deprecated

Returns the monitor number in which the point (x,y) is located.

Deprecated since 3.22

Use Display::get_monitor_at_point instead

x

the x coordinate in the virtual screen.

y

the y coordinate in the virtual screen.

Returns

the monitor number in which the point (x,y) lies, or a monitor close to (x,y) if the point is not in any monitor.

pub fn get_monitor_at_window(&self, window: &Window) -> i32

👎 Deprecated

Returns the number of the monitor in which the largest area of the bounding rectangle of window resides.

Deprecated since 3.22

Use Display::get_monitor_at_window instead

window

a Window

Returns

the monitor number in which most of window is located, or if window does not intersect any monitors, a monitor, close to window.

pub fn get_monitor_geometry(&self, monitor_num: i32) -> Rectangle

👎 Deprecated

Retrieves the Rectangle representing the size and position of the individual monitor within the entire screen area. The returned geometry is in ”application pixels”, not in ”device pixels” (see Screen::get_monitor_scale_factor).

Monitor numbers start at 0. To obtain the number of monitors of self, use Screen::get_n_monitors.

Note that the size of the entire screen area can be retrieved via Screen::get_width and Screen::get_height.

Deprecated since 3.22

Use Monitor::get_geometry instead

monitor_num

the monitor number

dest

a Rectangle to be filled with the monitor geometry

pub fn get_monitor_height_mm(&self, monitor_num: i32) -> i32

👎 Deprecated

Gets the height in millimeters of the specified monitor.

Deprecated since 3.22

Use Monitor::get_height_mm instead

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the height of the monitor, or -1 if not available

pub fn get_monitor_plug_name(&self, monitor_num: i32) -> Option<GString>

👎 Deprecated

Returns the output name of the specified monitor. Usually something like VGA, DVI, or TV, not the actual product name of the display device.

Deprecated since 3.22

Use Monitor::get_model instead

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

a newly-allocated string containing the name of the monitor, or None if the name cannot be determined

pub fn get_monitor_scale_factor(&self, monitor_num: i32) -> i32

👎 Deprecated

Returns the internal scale factor that maps from monitor coordinates to the actual device pixels. On traditional systems this is 1, but on very high density outputs this can be a higher value (often 2).

This can be used if you want to create pixel based data for a particular monitor, but most of the time you’re drawing to a window where it is better to use Window::get_scale_factor instead.

Deprecated since 3.22

Use Monitor::get_scale_factor instead

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the scale factor

pub fn get_monitor_width_mm(&self, monitor_num: i32) -> i32

👎 Deprecated

Gets the width in millimeters of the specified monitor, if available.

Deprecated since 3.22

Use Monitor::get_width_mm instead

monitor_num

number of the monitor, between 0 and gdk_screen_get_n_monitors (screen)

Returns

the width of the monitor, or -1 if not available

pub fn get_monitor_workarea(&self, monitor_num: i32) -> Rectangle

👎 Deprecated

Retrieves the Rectangle representing the size and position of the “work area” on a monitor within the entire screen area. The returned geometry is in ”application pixels”, not in ”device pixels” (see Screen::get_monitor_scale_factor).

The work area should be considered when positioning menus and similar popups, to avoid placing them below panels, docks or other desktop components.

Note that not all backends may have a concept of workarea. This function will return the monitor geometry if a workarea is not available, or does not apply.

Monitor numbers start at 0. To obtain the number of monitors of self, use Screen::get_n_monitors.

Deprecated since 3.22

Use Monitor::get_workarea instead

monitor_num

the monitor number

dest

a Rectangle to be filled with the monitor workarea

pub fn get_n_monitors(&self) -> i32

👎 Deprecated

Returns the number of monitors which self consists of.

Deprecated since 3.22

Use Display::get_n_monitors instead

Returns

number of monitors which self consists of

pub fn get_number(&self) -> i32

👎 Deprecated

Gets the index of self among the screens in the display to which it belongs. (See Screen::get_display)

Deprecated since 3.22

Returns

the index

pub fn get_primary_monitor(&self) -> i32

👎 Deprecated

Gets the primary monitor for self. The primary monitor is considered the monitor where the “main desktop” lives. While normal application windows typically allow the window manager to place the windows, specialized desktop applications such as panels should place themselves on the primary monitor.

If no primary monitor is configured by the user, the return value will be 0, defaulting to the first monitor.

Deprecated since 3.22

Use Display::get_primary_monitor instead

Returns

An integer index for the primary monitor, or 0 if none is configured.

pub fn get_resolution(&self) -> f64

Gets the resolution for font handling on the screen; see Screen::set_resolution for full details.

Returns

the current resolution, or -1 if no resolution has been set.

pub fn get_rgba_visual(&self) -> Option<Visual>

Gets a visual to use for creating windows with an alpha channel. The windowing system on which GTK+ is running may not support this capability, in which case None will be returned. Even if a non-None value is returned, its possible that the window’s alpha channel won’t be honored when displaying the window on the screen: in particular, for X an appropriate windowing manager and compositing manager must be running to provide appropriate display.

This functionality is not implemented in the Windows backend.

For setting an overall opacity for a top-level window, see Window::set_opacity.

Returns

a visual to use for windows with an alpha channel or None if the capability is not available.

pub fn get_root_window(&self) -> Option<Window>

Gets the root window of self.

Returns

the root window

pub fn get_system_visual(&self) -> Option<Visual>

Get the system’s default visual for self. This is the visual for the root window of the display. The return value should not be freed.

Returns

the system visual

pub fn get_toplevel_windows(&self) -> Vec<Window>

Obtains a list of all toplevel windows known to GDK on the screen self. A toplevel window is a child of the root window (see gdk_get_default_root_window).

The returned list should be freed with glib::List::free, but its elements need not be freed.

Returns

list of toplevel windows, free with glib::List::free

pub fn get_width(&self) -> i32

👎 Deprecated

Gets the width of self in pixels. The returned size is in ”application pixels”, not in ”device pixels” (see Screen::get_monitor_scale_factor).

Deprecated since 3.22

Use per-monitor information instead

Returns

the width of self in pixels.

pub fn get_width_mm(&self) -> i32

👎 Deprecated

Gets the width of self in millimeters.

Note that this value is somewhat ill-defined when the screen has multiple monitors of different resolution. It is recommended to use the monitor dimensions instead.

Deprecated since 3.22

Use per-monitor information instead

Returns

the width of self in millimeters.

pub fn get_window_stack(&self) -> Vec<Window>

Returns a glib::List of GdkWindows representing the current window stack.

On X11, this is done by inspecting the _NET_CLIENT_LIST_STACKING property on the root window, as described in the Extended Window Manager Hints. If the window manager does not support the _NET_CLIENT_LIST_STACKING hint, this function returns None.

On other platforms, this function may return None, depending on whether it is implementable on that platform.

The returned list is newly allocated and owns references to the windows it contains, so it should be freed using glib::List::free and its windows unrefed using gobject::Object::unref when no longer needed.

Returns

a list of GdkWindows for the current window stack, or None.

pub fn is_composited(&self) -> bool

Returns whether windows with an RGBA visual can reasonably be expected to have their alpha channel drawn correctly on the screen.

On X11 this function returns whether a compositing manager is compositing self.

Returns

Whether windows with RGBA visuals can reasonably be expected to have their alpha channels drawn correctly on the screen.

pub fn list_visuals(&self) -> Vec<Visual>

Lists the available visuals for the specified self. A visual describes a hardware image data format. For example, a visual might support 24-bit color, or 8-bit color, and might expect pixels to be in a certain format.

Call glib::List::free on the return value when you’re finished with it.

Returns

a list of visuals; the list must be freed, but not its contents

pub fn make_display_name(&self) -> GString

👎 Deprecated

Determines the name to pass to Display::open to get a Display with this screen as the default screen.

Deprecated since 3.22

Returns

a newly allocated string, free with g_free

pub fn set_font_options(&self, options: Option<&FontOptions>)

Sets the default font options for the screen. These options will be set on any pango::Context’s newly created with gdk_pango_context_get_for_screen. Changing the default set of font options does not affect contexts that have already been created.

options

a cairo::FontOptions, or None to unset any previously set default font options.

pub fn set_resolution(&self, dpi: f64)

Sets the resolution for font handling on the screen. This is a scale factor between points specified in a pango::FontDescription and cairo units. The default value is 96, meaning that a 10 point font will be 13 units high. (10 * 96. / 72. = 13.3).

dpi

the resolution in “dots per inch”. (Physical inches aren’t actually involved; the terminology is conventional.)

pub fn get_default() -> Option<Screen>

Gets the default screen for the default display. (See gdk_display_get_default ()).

Returns

a Screen, or None if there is no default display.

pub fn height() -> i32

👎 Deprecated

Gets the height of the default screen in pixels. The returned size is in ”application pixels”, not in ”device pixels” (see Screen::get_monitor_scale_factor).

Deprecated since 3.22

Use per-monitor information

Returns

the height of the default screen in pixels.

pub fn height_mm() -> i32

👎 Deprecated

Returns the height of the default screen in millimeters. Note that on many X servers this value will not be correct.

Deprecated since 3.22

Use per-monitor information

Returns

the height of the default screen in millimeters, though it is not always correct.

pub fn width() -> i32

👎 Deprecated

Gets the width of the default screen in pixels. The returned size is in ”application pixels”, not in ”device pixels” (see Screen::get_monitor_scale_factor).

Deprecated since 3.22

Use per-monitor information

Returns

the width of the default screen in pixels.

pub fn width_mm() -> i32

👎 Deprecated

Returns the width of the default screen in millimeters. Note that on many X servers this value will not be correct.

Deprecated since 3.22

Use per-monitor information

Returns

the width of the default screen in millimeters, though it is not always correct.

pub fn connect_composited_changed<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::composited-changed signal is emitted when the composited status of the screen changes

pub fn connect_monitors_changed<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::monitors-changed signal is emitted when the number, size or position of the monitors attached to the screen change.

Only for X11 and OS X for now. A future implementation for Win32 may be a possibility.

pub fn connect_size_changed<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

The ::size-changed signal is emitted when the pixel width or height of a screen changes.

pub fn connect_property_font_options_notify<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

pub fn connect_property_resolution_notify<F: Fn(&Screen) + 'static>(
&self,
f: F
) -> SignalHandlerId

impl Screen

pub fn get_font_options(&self) -> Option<FontOptions>

Gets any options previously set with Screen::set_font_options.

Returns

the current font options, or None if no default font options have been set.

pub fn get_setting(&self, name: &str) -> Option<Value>

Retrieves a desktop-wide setting such as double-click time for the Screen self.

FIXME needs a list of valid settings here, or a link to more information.

name

the name of the setting

value

location to store the value of the setting

Returns

true if the setting existed and a value was stored in value, false otherwise.