Struct gdk::Visual

pub struct Visual(_, _);

A Visual contains information about a particular visual.

Methods

impl Visual

fn get_bits_per_rgb(&self) -> i32

Returns the number of significant bits per red, green and blue value.

Returns

The number of significant bits per color value for self.

fn get_blue_pixel_details(&self) -> (u32, i32, i32)

Obtains values that are needed to calculate blue pixel values in TrueColor and DirectColor. The “mask” is the significant bits within the pixel. The “shift” is the number of bits left we must shift a primary for it to be in position (according to the "mask"). Finally, "precision" refers to how much precision the pixel value contains for a particular primary.

mask

A pointer to a guint32 to be filled in, or None

shift

A pointer to a gint to be filled in, or None

precision

A pointer to a gint to be filled in, or None

fn get_byte_order(&self) -> ByteOrder

Returns the byte order of this visual.

Returns

A ByteOrder stating the byte order of self.

fn get_colormap_size(&self) -> i32

Returns the size of a colormap for this visual.

Returns

The size of a colormap that is suitable for self.

fn get_depth(&self) -> i32

Returns the bit depth of this visual.

Returns

The bit depth of this visual.

fn get_green_pixel_details(&self) -> (u32, i32, i32)

Obtains values that are needed to calculate green pixel values in TrueColor and DirectColor. The “mask” is the significant bits within the pixel. The “shift” is the number of bits left we must shift a primary for it to be in position (according to the "mask"). Finally, "precision" refers to how much precision the pixel value contains for a particular primary.

mask

A pointer to a guint32 to be filled in, or None

shift

A pointer to a gint to be filled in, or None

precision

A pointer to a gint to be filled in, or None

fn get_red_pixel_details(&self) -> (u32, i32, i32)

Obtains values that are needed to calculate red pixel values in TrueColor and DirectColor. The “mask” is the significant bits within the pixel. The “shift” is the number of bits left we must shift a primary for it to be in position (according to the "mask"). Finally, "precision" refers to how much precision the pixel value contains for a particular primary.

mask

A pointer to a guint32 to be filled in, or None

shift

A pointer to a gint to be filled in, or None

precision

A pointer to a gint to be filled in, or None

fn get_screen(&self) -> Screen

Gets the screen to which this visual belongs

Returns

the screen to which this visual belongs.

fn get_visual_type(&self) -> VisualType

Returns the type of visual this is (PseudoColor, TrueColor, etc).

Returns

A VisualType stating the type of self.

fn get_best() -> Visual

Get the visual with the most available colors for the default GDK screen. The return value should not be freed.

Returns

best visual

fn get_best_depth() -> i32

Get the best available depth for the default GDK screen. “Best” means “largest,” i.e. 32 preferred over 24 preferred over 8 bits per pixel.

Returns

best available depth

fn get_best_type() -> VisualType

Return the best available visual type for the default GDK screen.

Returns

best visual type

fn get_best_with_both(depth: i32, visual_type: VisualType) -> Option<Visual>

Combines Visual::get_best_with_depth and Visual::get_best_with_type.

depth

a bit depth

visual_type

a visual type

Returns

best visual with both depth and visual_type, or None if none

fn get_best_with_depth(depth: i32) -> Option<Visual>

Get the best visual with depth depth for the default GDK screen. Color visuals and visuals with mutable colormaps are preferred over grayscale or fixed-colormap visuals. The return value should not be freed. None may be returned if no visual supports depth.

depth

a bit depth

Returns

best visual for the given depth

fn get_best_with_type(visual_type: VisualType) -> Option<Visual>

Get the best visual of the given visual_type for the default GDK screen. Visuals with higher color depths are considered better. The return value should not be freed. None may be returned if no visual has type visual_type.

visual_type

a visual type

Returns

best visual of the given type

fn get_system() -> Visual

Get the system’s default visual for the default GDK screen. This is the visual for the root window of the display. The return value should not be freed.

Returns

system visual

impl Visual

fn query_depths() -> Vec<i32>