Struct gdk::Visual

pub struct Visual(_, _);

A Visual contains information about a particular visual.

Implements

VisualExt

Methods

impl Visual

pub fn get_best() -> Visual

Get the visual with the most available colors for the default GDK screen. The return value should not be freed.

Deprecated since 3.22

Visual selection should be done using ScreenExt::get_system_visual and ScreenExt::get_rgba_visual

Returns

best visual

pub fn get_best_depth() -> i32

Get the best available depth for the default GDK screen. “Best” means “largest,” i.e. 32 preferred over 24 preferred over 8 bits per pixel.

Deprecated since 3.22

Visual selection should be done using ScreenExt::get_system_visual and ScreenExt::get_rgba_visual

Returns

best available depth

pub fn get_best_type() -> VisualType

Return the best available visual type for the default GDK screen.

Deprecated since 3.22

Visual selection should be done using ScreenExt::get_system_visual and ScreenExt::get_rgba_visual

Returns

best visual type

pub fn get_best_with_both(depth: i32, visual_type: VisualType) -> Option<Visual>

Combines Visual::get_best_with_depth and Visual::get_best_with_type.

Deprecated since 3.22

Visual selection should be done using ScreenExt::get_system_visual and ScreenExt::get_rgba_visual

depth

a bit depth

visual_type

a visual type

Returns

best visual with both depth and visual_type, or None if none

pub fn get_best_with_depth(depth: i32) -> Option<Visual>

Get the best visual with depth depth for the default GDK screen. Color visuals and visuals with mutable colormaps are preferred over grayscale or fixed-colormap visuals. The return value should not be freed. None may be returned if no visual supports depth.

Deprecated since 3.22

Visual selection should be done using ScreenExt::get_system_visual and ScreenExt::get_rgba_visual

depth

a bit depth

Returns

best visual for the given depth

pub fn get_best_with_type(visual_type: VisualType) -> Option<Visual>

Get the best visual of the given visual_type for the default GDK screen. Visuals with higher color depths are considered better. The return value should not be freed. None may be returned if no visual has type visual_type.

Deprecated since 3.22

Visual selection should be done using ScreenExt::get_system_visual and ScreenExt::get_rgba_visual

visual_type

a visual type

Returns

best visual of the given type

pub fn get_system() -> Visual

Get the system’s default visual for the default GDK screen. This is the visual for the root window of the display. The return value should not be freed.

Deprecated since 3.22

Use gdk_screen_get_system_visual (gdk_screen_get_default ()).

Returns

system visual