Struct gdk::Window

pub struct Window(_, _);

Implements

WindowExt

Methods

impl Window

pub fn at_pointer() -> (Window, i32, i32)

Deprecated

Obtains the window underneath the mouse pointer, returning the location of that window in win_x, win_y. Returns None if the window under the mouse pointer is not known to GDK (if the window belongs to another application and a Window hasn’t been created for it with gdk_window_foreign_new)

NOTE: For multihead-aware widgets or applications use DisplayExt::get_window_at_pointer instead.

Deprecated since 3.0

Use DeviceExt::get_window_at_position instead.

win_x

return location for origin of the window under the pointer

win_y

return location for origin of the window under the pointer

Returns

window under the mouse pointer

pub fn constrain_size(
geometry: &mut Geometry,
flags: WindowHints,
width: i32,
height: i32
) -> (i32, i32)

Constrains a desired width and height according to a set of geometry hints (such as minimum and maximum size).

geometry

a Geometry structure

flags

a mask indicating what portions of geometry are set

width

desired width of window

height

desired height of the window

new_width

location to store resulting width

new_height

location to store resulting height

pub fn process_all_updates()

Calls WindowExt::process_updates for all windows (see Window) in the application.

Deprecated since 3.22

pub fn set_debug_updates(setting: bool)

With update debugging enabled, calls to WindowExt::invalidate_region clear the invalidated region of the screen to a noticeable color, and GDK pauses for a short time before sending exposes to windows during WindowExt::process_updates. The net effect is that you can see the invalid region for each window and watch redraws as they occur. This allows you to diagnose inefficiencies in your application.

In essence, because the GDK rendering model prevents all flicker, if you are redrawing the same region 400 times you may never notice, aside from noticing a speed problem. Enabling update debugging causes GTK to flicker slowly and noticeably, so you can see exactly what’s being redrawn when, in what order.

The --gtk-debug=updates command line option passed to GTK+ programs enables this debug option at application startup time. That's usually more useful than calling Window::set_debug_updates yourself, though you might want to use this function to enable updates sometime after application startup time.

Deprecated since 3.22

setting

true to turn on update debugging

impl Window

pub fn new(parent: Option<&Window>, attributes: &WindowAttr) -> Window

Creates a new Window using the attributes from attributes. See WindowAttr and WindowAttributesType for more details. Note: to use this on displays other than the default display, parent must be specified.

parent

a Window, or None to create the window as a child of the default root window for the default display.

attributes

attributes of the new window

attributes_mask

mask indicating which fields in attributes are valid

Returns

the new Window

pub fn create_similar_surface(
&self,
content: Content,
width: i32,
height: i32
) -> Option<Surface>